% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{add_chMDB_user}
\alias{add_chMDB_user}
\title{Add a user to an MDB of a \link{chTKCat} object}
\usage{
add_chMDB_user(x, mdb, login, admin = FALSE)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{mdb}{name of the modeled database}

\item{login}{login of the user to drop}

\item{admin}{if the user is an admin of the MDB}
}
\value{
No return value, called for side effects
}
\description{
Add a user to an MDB of a \link{chTKCat} object
}
