% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDB.R
\name{add_collection_member}
\alias{add_collection_member}
\title{Add a collection member to an MDB}
\usage{
add_collection_member(x, collection, table, ...)
}
\arguments{
\item{x}{an \link{MDB} object}

\item{collection}{a collection title in \code{\link[=list_local_collections]{list_local_collections()}}}

\item{table}{the table providing the collection member}

\item{...}{definition of the collection fields as lists
(e.g. \code{be=list(static=TRUE, value="Gene")}
or
\code{organism=list(static=TRUE, value="Homo sapiens", type="Scientific name")}
)}
}
\description{
Add a collection member to an MDB
}
