% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{add_feature_def}
\alias{add_feature_def}
\title{Add a feature definition to Knowledge Management Requirements (KMR)}
\usage{
add_feature_def(kmr, name, description, properties)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{name}{the name of the feature type}

\item{description}{description of the feature type}

\item{properties}{properties of the feature.
A list named with property names. For each property, a list with:
\itemize{
\item \strong{type}: among "integer", "numeric", "logical", "character", "Date",
"POSIXct", "base64", "table" and "field" ("table" and "field" type are used
for referencing tables and fields; the other types come from ReDaMoR).
\item \strong{description} (optional): a description of the feature property. Useful
when the feature has more than one property
\item \strong{mandatory}: a logical indicating the property is mandatory for the
feature
\item \strong{measurement} (optional and only for "integer" and "numeric" types):
the name of the measurement for checking units
}}
}
\value{
The modified KMR \link{MDB} object
}
\description{
Add a feature definition to Knowledge Management Requirements (KMR)
}
