% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_kmspec.R
\name{add_helpers}
\alias{add_helpers}
\title{Add helpers in an \link{MDB} object}
\usage{
add_helpers(x, kmr, code, name = "R-Helpers", language = "R")
}
\arguments{
\item{x}{the MDB object}

\item{kmr}{an \link{MDB} object with KM requirements}

\item{code}{file path to the source code of helper functions}

\item{name}{the name of the helper set (default: "R-Helpers")}

\item{language}{the programing language of the code (default: "R")}
}
\value{
An \link{MDB} object with additional helpers
}
\description{
Add helpers in an \link{MDB} object
}
