% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM-specifications.R
\name{add_km_spec}
\alias{add_km_spec}
\title{Add empty KM specification tables to an \link{MDB} object}
\usage{
add_km_spec(x, kmr)
}
\arguments{
\item{x}{an \link{MDB} object to update with specification tables}

\item{kmr}{an \link{MDB} object with KM requirements}
}
\value{
An \link{MDB} object with empty KM specification tables
}
\description{
Add empty KM specification tables to an \link{MDB} object
}
