% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM-specifications.R
\name{add_km_table}
\alias{add_km_table}
\title{Add KM table specifications to an \link{MDB} object}
\usage{
add_km_table(x, kmr, name, type, features = list())
}
\arguments{
\item{x}{an \link{MDB} object to update with specification tables}

\item{kmr}{an \link{MDB} object with KM requirements}

\item{name}{the name of an existing table in x}

\item{type}{the name of an existing table type in kmr}

\item{features}{a list of feature definitions. Each element of a list
is a list of parameters for the \code{\link[=add_km_feature]{add_km_feature()}} function.}
}
\value{
An \link{MDB} object with additional KM table specification
}
\description{
Add KM table specifications to an \link{MDB} object
}
