% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{add_property_values}
\alias{add_property_values}
\title{Add possible values to an integer or a character feature property in \link{KMR}}
\usage{
add_property_values(kmr, feature, property, values)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{feature}{the name of the feature type}

\item{property}{the name of the property}

\item{values}{a vector of character or integer or a named vector with the
description of the values}
}
\value{
The modified KMR \link{MDB} object
}
\description{
Add possible values to an integer or a character feature property in \link{KMR}
}
