% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{add_table_def}
\alias{add_table_def}
\title{Add a table definition to Knowledge Management Requirements (KMR)}
\usage{
add_table_def(
  kmr,
  name,
  description,
  collection = as.character(NA),
  mandatory_features
)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{name}{the name of the table type}

\item{description}{description of the table type}

\item{collection}{the name of the collection of which this table type
must be a member (default: NA)}

\item{mandatory_features}{a character vector with mandatory features for this
table type}
}
\value{
The modified KMR \link{MDB} object
}
\description{
Add a table definition to Knowledge Management Requirements (KMR)
}
