% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{add_table_features}
\alias{add_table_features}
\title{Add possible features to table type in \link{KMR}}
\usage{
add_table_features(kmr, table, features)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{table}{the name of the table type}

\item{features}{a character vector with optional features for this table type}
}
\value{
The modified KMR \link{MDB} object
}
\description{
Add possible features to table type in \link{KMR}
}
