% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{add_unit_def}
\alias{add_unit_def}
\title{Add a unit definition to Knowledge Management Requirements (KMR)}
\usage{
add_unit_def(kmr, measurement, unit, description)
}
\arguments{
\item{kmr}{an \link{KMR} object}

\item{measurement}{the name of the measurement (e.g. "volume")}

\item{unit}{the name of the unit (e.g. "ml")}

\item{description}{a description of the unit (e.g. "milliliter)}
}
\value{
The modified \link{KMR} object
}
\description{
Add a unit definition to Knowledge Management Requirements (KMR)
}
