% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMDBToChTKCat.R
\name{addMDBToChTKCat}
\alias{addMDBToChTKCat}
\title{Add an MDB (modeled database) to a \link{chTKCat} database}
\usage{
addMDBToChTKCat(tkcon, imdb)
}
\arguments{
\item{tkcon}{a \link{chTKCat} object}

\item{imdb}{an \link{internalMDB} object}
}
\description{
Add an MDB (modeled database) to a \link{chTKCat} database
}
