% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{archive_chMDB}
\alias{archive_chMDB}
\title{Archive a chMDB in a \link{chTKCat}}
\usage{
archive_chMDB(
  x,
  name,
  defaultTS = as.POSIXct("1970-01-01 00:00.0", tz = "UTC")
)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the database to archive}

\item{defaultTS}{a default timestamp value to use when not existing in
the DB (default: \code{as.POSIXct("1970-01-01 00:00.0", tz="UTC")})}
}
\value{
No return value, called for side effects
}
\description{
Archive a chMDB in a \link{chTKCat}
}
