% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{as.internalMDB}
\alias{as.internalMDB}
\title{Convert an object in an \link{internalMDB} object}
\usage{
as.internalMDB(x, checks = c())
}
\arguments{
\item{x}{any object with relevant \code{dataModel}, \code{dataTables}, \code{dbInfo}
and \code{collectionMembers} methods}
}
\value{
An \link{internalMDB} object
}
\description{
Convert an object in an \link{internalMDB} object
}
