% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/memoMDB.R
\name{as_chMDB}
\alias{as_chMDB}
\alias{as_chMDB.memoMDB}
\title{Push an MDB object in a ClickHouse database}
\usage{
as_chMDB(x, tkcon, overwrite = FALSE, ...)

\method{as_chMDB}{memoMDB}(x, tkcon, overwrite = FALSE, ...)
}
\arguments{
\item{x}{an MDB object}

\item{tkcon}{a \link{chTKCat} object}

\item{overwrite}{a logical indicating if existing data should be overwritten
(default: FALSE)}

\item{...}{method specific parameters}
}
\value{
A \link{chMDB} object.
}
\description{
Push an MDB object in a ClickHouse database
}
