% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/fileMDB.R, R/generics.R,
%   R/memoMDB.R, R/metaMDB.R
\name{as_fileMDB.chMDB}
\alias{as_fileMDB.chMDB}
\alias{as_fileMDB.fileMDB}
\alias{as_fileMDB}
\alias{as_fileMDB.memoMDB}
\alias{as_fileMDB.metaMDB}
\title{Write an MDB object}
\usage{
\method{as_fileMDB}{chMDB}(x, path, by = 10^5, ...)

\method{as_fileMDB}{fileMDB}(x, path, ...)

as_fileMDB(x, path, ...)

\method{as_fileMDB}{memoMDB}(x, path, ...)

\method{as_fileMDB}{metaMDB}(x, path, ...)
}
\arguments{
\item{x}{an MDB object}

\item{path}{the path where the MDB should be written}

\item{by}{the size of the batch: number of records to write
together (default: 10^5)}

\item{...}{method specific parameters}
}
\value{
A \link{fileMDB} object.
}
\description{
Write an MDB object
}
