% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{as_KMR}
\alias{as_KMR}
\title{Convert in a \link{KMR} object when possible}
\usage{
as_KMR(x)
}
\arguments{
\item{x}{an object}
}
\value{
A \link{KMR} object
}
\description{
Convert in a \link{KMR} object when possible
}
