% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memeMDB.R
\name{as_memeMDB}
\alias{as_memeMDB}
\title{Convert  any MDB object in a \link{memeMDB} object}
\usage{
as_memeMDB(x, ...)
}
\arguments{
\item{x}{a MDB object}

\item{...}{additional parameters for the \code{\link[=memeMDB]{memeMDB()}} function.}
}
\value{
A \link{memeMDB} object
}
\description{
Convert  any MDB object in a \link{memeMDB} object
}
