% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoMDB.R
\name{as_memoMDB}
\alias{as_memoMDB}
\title{Convert  any MDB object in a \link{memoMDB} object}
\usage{
as_memoMDB(x, ...)
}
\arguments{
\item{x}{a MDB object}

\item{...}{additional parameters for the \code{\link[=memoMDB]{memoMDB()}} function.}
}
\value{
A \link{memoMDB} object
}
\description{
Convert  any MDB object in a \link{memoMDB} object
}
