% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R
\name{ch_insert}
\alias{ch_insert}
\title{Insert records by batches in a Clickhouse table}
\usage{
ch_insert(con, dbName, tableName, value, by = 10^6)
}
\arguments{
\item{con}{the clickhouse connection}

\item{dbName}{the name of the database}

\item{tableName}{the name of the table}

\item{value}{the table to import}

\item{by}{the size of the batch: number of records to import
together (default: 10^6)}
}
\value{
No return value, called for side effects
}
\description{
Insert records by batches in a Clickhouse table
}
