% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R
\name{ch_list_tables}
\alias{ch_list_tables}
\title{List tables from a ClickHouse database}
\usage{
ch_list_tables(con, dbName = NA)
}
\arguments{
\item{con}{the clickhouse connection}

\item{dbName}{the name of the database from which the tables should be listed}
}
\value{
A character vector with table names
}
\description{
List tables from a ClickHouse database
}
