% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R
\name{ch_reconnect}
\alias{ch_reconnect}
\title{Reconnect to a ClickHouse database}
\usage{
ch_reconnect(x, user, password, ntries = 3)
}
\arguments{
\item{x}{a ClickhouseConnection object}

\item{user}{user name. If not provided, it's taken from x}

\item{password}{user password. If not provided, first the function
tries to connect without any password.If it fails, the function asks the
user to provide a password.}

\item{ntries}{the number of times the user can enter a wrong password}
}
\description{
Reconnect to a ClickHouse database
}
