% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{change_chTKCat_password}
\alias{change_chTKCat_password}
\title{Change chTKCat password}
\usage{
change_chTKCat_password(x, login, password)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{login}{user login}

\item{password}{new user password}
}
\value{
No return value, called for side effects
}
\description{
Change chTKCat password
}
