% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{check_chTKCat}
\alias{check_chTKCat}
\title{Check the chTKCat object}
\usage{
check_chTKCat(x, verbose = FALSE)
}
\arguments{
\item{x}{a chTKCat object}

\item{verbose}{a logical indicating if information messages should be
displayed.}
}
\value{
Invisible result: chTKCat object
}
\description{
Check the chTKCat object
}
\seealso{
\code{\link{connectToTbkm}}
}
