% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_kmspec.R
\name{check_km_spec}
\alias{check_km_spec}
\title{Check if data shared in an \link{MDB} object fits the KM requirements}
\usage{
check_km_spec(x, kmr)
}
\arguments{
\item{x}{an \link{MDB} object to check}

\item{kmr}{an \link{MDB} object with KM requirements}
}
\value{
A logical: TRUE if requirements are met, FALSE if not.
report is also generated indicating error or warnings.
}
\description{
Check if data shared in an \link{MDB} object fits the KM requirements
}
