% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{checkTable}
\alias{checkTable}
\title{Check table according to a data model}
\usage{
checkTable(x, tableModel)
}
\arguments{
\item{x}{a data.frame}

\item{tableModel}{a \code{\link{RelTableModel}} object}
}
\description{
Check table according to a data model
}
