% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{collection_members<-}
\alias{collection_members<-}
\title{Set collection members}
\usage{
collection_members(x) <- value
}
\arguments{
\item{x}{an object with embedded collection members}

\item{value}{the new collection members}
}
\description{
Set collection members
}
