% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{collection_members.chTKCat}
\alias{collection_members.chTKCat}
\title{Get collection members of a \link{chTKCat} object}
\usage{
\method{collection_members}{chTKCat}(x, collections = NULL, ...)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{collections}{a character vector indicating the name of the collections
to focus on (default: NULL ==> all of them)}

\item{...}{not used}
}
\description{
Get collection members of a \link{chTKCat} object
}
