% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{collection_members}
\alias{collection_members}
\title{Get collection members}
\usage{
collection_members(x, ...)
}
\arguments{
\item{x}{an object with embedded collection members}

\item{...}{further arguments for the method}
}
\description{
Get collection members
}
