% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{collectionMembers<-.internalMDB}
\alias{collectionMembers<-.internalMDB}
\title{Set collection members of an \link{internalMDB} object}
\usage{
\method{collectionMembers}{internalMDB}(x) <- value
}
\arguments{
\item{x}{an \link{internalMDB} object}

\item{value}{A data.frame with the following columns:
\itemize{
\item \strong{collection} (character): The name of the collection
\item \strong{resource} (character): The name of the resource
\item \strong{cid} (integer): collection member ID by resource
\item \strong{table} (character): The table recording collection information
\item \strong{field} (character): The collection field.
\item \strong{static} (logical): TRUE if the field value is common to all elements.
\item \strong{value} (character): The name of the table column if static is FALSE
or the field value if static is TRUE.
\item \strong{type} (character): the type of the field.
(not necessarily used ==> NA if not)
}}
}
\description{
Set collection members of an \link{internalMDB} object
}
