% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{collectionMembers.chMDB}
\alias{collectionMembers.chMDB}
\title{Get collection members of a \link{chMDB} object}
\usage{
\method{collectionMembers}{chMDB}(x, collections = NULL)
}
\arguments{
\item{x}{a \link{chMDB} object}

\item{collection}{a character vector indicating the name of the collections
to focus on (default: NULL ==> all of them)}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with the following columns:
\itemize{
\item \strong{collection} (character): The name of the collection
\item \strong{resource} (character): The name of the resource
\item \strong{cid} (integer): collection member ID by resource
\item \strong{table} (character): The table recording collection information
\item \strong{field} (character): The collection field.
\item \strong{static} (logical): TRUE if the field value is common to all elements.
\item \strong{value} (character): The name of the table column if static is FALSE
or the field value if static is TRUE.
\item \strong{type} (character): the type of the field.
(not necessarily used ==> NA if not)
}
}
\description{
Get collection members of a \link{chMDB} object
}
