% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_MDB.R
\name{compare_MDB}
\alias{compare_MDB}
\title{Compare two MDB}
\usage{
compare_MDB(former, new)
}
\arguments{
\item{former}{an MDB object (e.g. \link{internalMDB} or \link{chMDB})}

\item{new}{an MDB object (e.g. \link{internalMDB} or \link{chMDB})}
}
\value{
A tibble with 3 columns:
\itemize{
\item \strong{Information}: Compared information
\item \strong{Former}: former value
\item \strong{New}: new value
\item \strong{Identical}: a logical indicating if the 2 values are identical
}
}
\description{
Compare two MDB
}
