% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memeMDB.R
\name{count_records.fileMDB}
\alias{count_records.fileMDB}
\alias{count_records}
\alias{count_records.memeMDB}
\title{Count the number of records}
\usage{
\method{count_records}{fileMDB}(x, ...)

count_records(x, ...)

\method{count_records}{memeMDB}(x, ...)
}
\arguments{
\item{x}{a \link{memeMDB}}

\item{...}{the name of the tables to consider (default: all of them)}
}
\value{
A named vector with the number of records per table.
}
\description{
Count the number of records
}
