% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{create_KMR}
\alias{create_KMR}
\alias{KMR}
\title{Create an \link{MDB} object with Knowledge Management Requirements (KMR)}
\usage{
create_KMR(name, title, description, version, maintainer)
}
\arguments{
\item{name}{the name of the requirements}

\item{title}{the title of the requirements}

\item{description}{the description of the requirements}

\item{version}{version of the requirements}

\item{maintainer}{maintainer of the requirements}
}
\value{
An MDB object with KM requirements data model
}
\description{
Create an \link{MDB} object with Knowledge Management Requirements (KMR)
}
