% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POK.R
\name{create_POK}
\alias{create_POK}
\alias{POK}
\title{Create a piece of knowledge (POK) from an \link{MDB} and a \link{KMR}object}
\usage{
create_POK(mdb, kmr, tkcat = NULL)
}
\arguments{
\item{mdb}{a \link{MDB} object with KM specifications}

\item{kmr}{a \link{KMR} object with KM requirements}

\item{tkcat}{A \link{TKCat} or \link{chTKCat} object to make available in
helper environment}
}
\value{
A POK object: a list with 3 slots:
\itemize{
\item $mdb: the provided \link{MDB} object
\item $kmr: the provided \link{KMR} object
\item $helpers: a list functions to leverage data from mdb and kmr
}
}
\description{
Create a piece of knowledge (POK) from an \link{MDB} and a \link{KMR}object
}
