% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{data_file_size}
\alias{data_file_size}
\title{Get the size of data files from a \link{fileMDB} object}
\usage{
data_file_size(x, hr = FALSE)
}
\arguments{
\item{x}{a \link{fileMDB} object}

\item{hr}{a logical indicating if the values should be "human readable".
(default: FALSE)}
}
\value{
a numeric vector with size in bytes (hr=FALSE) or
a character vector with size and units (hr=TRUE)
}
\description{
Get the size of data files from a \link{fileMDB} object
}
