% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{data_files}
\alias{data_files}
\title{Get the data files from an object}
\usage{
data_files(x)
}
\arguments{
\item{x}{an object with a dataFiles and readParameters slots}
}
\value{
a list with "dataFiles" and "readParameters" for reading the files.
}
\description{
Get the data files from an object
}
