% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{data_model.fileMDB}
\alias{data_model.fileMDB}
\title{Get the \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} of a \link{fileMDB} object}
\usage{
\method{data_model}{fileMDB}(x, ...)
}
\arguments{
\item{x}{a \link{fileMDB} object}

\item{...}{not used}
}
\value{
A \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object
}
\description{
Get the \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} of a \link{fileMDB} object
}
