% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R
\name{data_model.fileMDB}
\alias{data_model.fileMDB}
\alias{data_model}
\title{Get object data model}
\usage{
\method{data_model}{fileMDB}(x, ...)

data_model(x, ...)
}
\arguments{
\item{x}{an object with an embedded data model}

\item{...}{not used}
}
\value{
A \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object
}
\description{
Get object data model
}
