% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{data_tables}
\alias{data_tables}
\title{Get object data tables}
\usage{
data_tables(x, ...)
}
\arguments{
\item{x}{an object with embedded data tables}

\item{...}{further arguments for the method}
}
\description{
Get object data tables
}
