% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{dataModel.chMDB}
\alias{dataModel.chMDB}
\title{Get the \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} of a \link{chMDB} object}
\usage{
\method{dataModel}{chMDB}(x)
}
\arguments{
\item{x}{a \link{chMDB} object}
}
\value{
A \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object
}
\description{
Get the \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} of a \link{chMDB} object
}
