% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{dataModel.internalMDB}
\alias{dataModel.internalMDB}
\title{Get the \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} of an \link{internalMDB} object}
\usage{
\method{dataModel}{internalMDB}(x)
}
\arguments{
\item{x}{an \link{internalMDB} object}
}
\value{
A \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object
}
\description{
Get the \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} of an \link{internalMDB} object
}
