% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{dataTables.chMDB}
\alias{dataTables.chMDB}
\title{Get data tables from a \link{chMDB} object}
\usage{
\method{dataTables}{chMDB}(x, ...)
}
\arguments{
\item{x}{a \link{chMDB} object}

\item{...}{the name of the tables to get (default: all of them)}
}
\value{
A list of \link[tibble:tibble]{tibble::tibble}
}
\description{
Get data tables from a \link{chMDB} object
}
