% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{dataTables.internalMDB}
\alias{dataTables.internalMDB}
\title{Get data tables from an \link{internalMDB} object}
\usage{
\method{dataTables}{internalMDB}(x, ...)
}
\arguments{
\item{x}{an \link{internalMDB} object}

\item{...}{the name of the tables to get (default: all of them)}
}
\value{
A list of \link[tibble:tibble]{tibble::tibble}
}
\description{
Get data tables from an \link{internalMDB} object
}
