% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/chTKCat.R, R/generics.R
\name{db_disconnect.chMDB}
\alias{db_disconnect.chMDB}
\alias{db_disconnect.chTKCat}
\alias{db_disconnect}
\title{Disconnect an object from a database}
\usage{
\method{db_disconnect}{chMDB}(x)

\method{db_disconnect}{chTKCat}(x)

db_disconnect(x)
}
\arguments{
\item{x}{an object with a database connection}
}
\value{
No return value, called for side effects
}
\description{
Disconnect an object from a database
}
