% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R
\name{db_info<-.fileMDB}
\alias{db_info<-.fileMDB}
\alias{db_info<-}
\title{Set DB information}
\usage{
\method{db_info}{fileMDB}(x) <- value

db_info(x) <- value
}
\arguments{
\item{x}{an object with embedded database information}

\item{value}{list with the following elements:
\itemize{
\item \strong{name}: a single character
\item \strong{title}: a single character
\item \strong{description}: a single character
\item \strong{url}: a single character
\item \strong{version}: a single character
\item \strong{maintainer}: a single character vector
\item \strong{size}: a numeric vector providing the size of the DB in bytes
}}
}
\description{
Set DB information
}
