% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{db_info.fileMDB}
\alias{db_info.fileMDB}
\title{Get DB information of a \link{fileMDB} object}
\usage{
\method{db_info}{fileMDB}(x, ...)
}
\arguments{
\item{x}{a \link{fileMDB} object}

\item{...}{not used}
}
\value{
A list with the following elements:
\itemize{
\item \strong{name}: a single character
\item \strong{title}: a single character
\item \strong{description}: a single character
\item \strong{url}: a single character
\item \strong{version}: a single character
\item \strong{maintainer}: a single character vector
\item \strong{size}: a numeric vector providing the size of the DB in bytes
}
}
\description{
Get DB information of a \link{fileMDB} object
}
