% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{dbInfo.chMDB}
\alias{dbInfo.chMDB}
\title{Get DB information of a \link{chMDB} object}
\usage{
\method{dbInfo}{chMDB}(x, countRecords = TRUE)
}
\arguments{
\item{x}{a \link{chMDB} object}

\item{countRecords}{a single logical. If TRUE (default), the number of
records is also returned.}
}
\value{
A list with the following elements:
\itemize{
\item \strong{name}: a single character
\item \strong{title}: a single character
\item \strong{description}: a single character
\item \strong{url}: a single character
\item \strong{version}: a single character
\item (\strong{records}): a single numeric
}
}
\description{
Get DB information of a \link{chMDB} object
}
