% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{dbInfo.internalMDB}
\alias{dbInfo.internalMDB}
\title{Get DB information of an \link{internalMDB} object}
\usage{
\method{dbInfo}{internalMDB}(x)
}
\arguments{
\item{x}{an \link{internalMDB} object}
}
\value{
A list with the following elements:
\itemize{
\item \strong{name}: a single character
\item \strong{title}: a single character
\item \strong{description}: a single character
\item \strong{url}: a single character
\item \strong{version}: a single character
}
}
\description{
Get DB information of an \link{internalMDB} object
}
