% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-management.R
\name{decode_bin}
\alias{decode_bin}
\title{Decode base64 string}
\usage{
decode_bin(text)
}
\arguments{
\item{text}{the base64 character vector to decode}
}
\value{
One decoded value (e.g. a raw vector corresponding to a binary file)
}
\description{
Decode base64 string
}
