% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/fileMDB.R, R/generics.R,
%   R/memoMDB.R, R/metaMDB.R
\name{dims.chMDB}
\alias{dims.chMDB}
\alias{dims.fileMDB}
\alias{dims}
\alias{dims.memoMDB}
\alias{dims.metaMDB}
\title{Detailed information about the format of the tables}
\usage{
\method{dims}{chMDB}(x, ...)

\method{dims}{fileMDB}(x, ...)

dims(x, ...)

\method{dims}{memoMDB}(x, ...)

\method{dims}{metaMDB}(x, ...)
}
\arguments{
\item{x}{an object with embedded data tables}

\item{...}{the name of the tables to consider (default: all of them)}
}
\value{
A tibble with one row for each considered table and the
following columns:
\itemize{
\item name: the name of the table
\item format: "table" or "matrix"
\item ncol: number of columns
\item nrow: number of rows
\item records: number of records (\code{nrow} for tables and \code{ncol*nrow} for matrices)
\item transposed: FALSE by default. TRUE only for matrices stored in a
transposed format.
}
}
\description{
Detailed information about the format of the tables
}
