% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMDBToChTKCat.R
\name{.dbinsert}
\alias{.dbinsert}
\title{Insert records by batches in a Clickhouse table}
\usage{
.dbinsert(conn, dbName, tableName, value, by = 10^6)
}
\arguments{
\item{conn}{the clickhouse connection}

\item{dbName}{the name of the database}

\item{tableName}{the name of the table}

\item{value}{the table to import}

\item{by}{the size of the batch: number of records to import
together (default: 10^6)}
}
\description{
Insert records by batches in a Clickhouse table
}
