% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-helpers.R
\name{.format_bytes}
\alias{.format_bytes}
\title{Format bytes numbers in human readable values}
\usage{
.format_bytes(bytes)
}
\arguments{
\item{bytes}{a vector of integers}
}
\value{
A vector of character with human readable size
}
\description{
Format bytes numbers in human readable values
}
