% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{drop_chMDB}
\alias{drop_chMDB}
\title{Drop a database from a \link{chTKCat}}
\usage{
drop_chMDB(x, name)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the database to remove}
}
\value{
No return value, called for side effects
}
\description{
Drop a database from a \link{chTKCat}
}
