% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{drop_chMDB_projections}
\alias{drop_chMDB_projections}
\title{Drop in ClickHouse any projection documented in the data model}
\usage{
drop_chMDB_projections(x)
}
\arguments{
\item{x}{a \link{chMDB} object}
}
\value{
Nothing: drop projections in ClickHouse.
}
\description{
Drop in ClickHouse any projection documented in the data model
}
