% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{drop_chTKCat_user}
\alias{drop_chTKCat_user}
\title{Drop a user from a \link{chTKCat} object}
\usage{
drop_chTKCat_user(x, login)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{login}{login of the user to drop}
}
\value{
No return value, called for side effects
}
\description{
Drop a user from a \link{chTKCat} object
}
