% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-management.R
\name{encode_bin}
\alias{encode_bin}
\title{Encode a binary file in a base64 string}
\usage{
encode_bin(what)
}
\arguments{
\item{what}{a file path or a raw vector}
}
\value{
A character vector of length 1 with the base64 encoded file
}
\description{
Encode a binary file in a base64 string
}
