% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R, R/generics.R
\name{explore_MDBs.chTKCat}
\alias{explore_MDBs.chTKCat}
\alias{explore_MDBs}
\title{Explore available \link{MDB} in a shiny web interface}
\usage{
\method{explore_MDBs}{chTKCat}(x, subSetSize = 100, host = x$chcon@host)

explore_MDBs(x, ...)
}
\arguments{
\item{x}{a \link{TKCat} related object (e.g. \link{chTKCat})}

\item{...}{method specific parameters}
}
\description{
Explore available \link{MDB} in a shiny web interface
}
