% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{export.internalMDB}
\alias{export.internalMDB}
\title{Export an \link{internalMDB} in a folder}
\usage{
\method{export}{internalMDB}(x, path, overwrite = FALSE)
}
\arguments{
\item{x}{the \link{internalMDB} to export}

\item{path}{the directory to create}

\item{overwrite}{a logical indicating if an existing directory should
be overwritten (default: FALSE)}
}
\description{
Export an \link{internalMDB} in a folder
}
