% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{filter.fileMDB}
\alias{filter.fileMDB}
\title{Filter a \link{fileMDB} object and return a \link{memoMDB}}
\usage{
\method{filter}{fileMDB}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{a \link{fileMDB} object}

\item{...}{each argument should have the name of one of the tables of the
\link{fileMDB} object and contain a simple logical expression involving
the names of the corresponding table.}

\item{.preserve}{not used}
}
\value{
a \link{memoMDB} object
}
\description{
Filter a \link{fileMDB} object and return a \link{memoMDB}
}
