% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memeMDB.R
\name{filter.memeMDB}
\alias{filter.memeMDB}
\title{Filter a \link{memeMDB} object}
\usage{
\method{filter}{memeMDB}(x, ...)
}
\arguments{
\item{x}{a \link{memeMDB} object}

\item{...}{each argument should have the name of one of the tables of the
\link{memeMDB} object and contain a simple logical expression involving
the names of the corresponding table.}
}
\value{
a filtered \link{memeMDB} object
}
\description{
Filter a \link{memeMDB} object
}
\examples{
filter(hpo, HPO_diseases=(db=="DECIPHER" & id=="1"))

}
