% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoMDB.R
\name{filter.memoMDB}
\alias{filter.memoMDB}
\title{Filter a \link{memoMDB} object}
\usage{
\method{filter}{memoMDB}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{a \link{memoMDB} object}

\item{...}{each argument should have the name of one of the tables of the
\link{memoMDB} object and contain a simple logical expression involving
the names of the corresponding table.}

\item{.preserve}{not used}
}
\value{
a filtered \link{memoMDB} object
}
\description{
Filter a \link{memoMDB} object
}
