% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMDB.R
\name{filter.metaMDB}
\alias{filter.metaMDB}
\title{Filter a \link{metaMDB} object}
\usage{
\method{filter}{metaMDB}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{a \link{metaMDB} object}

\item{...}{each argument should have the name of one of the tables of the
\link{metaMDB} object and contain a simple logical expression involving
the names of the corresponding table.}

\item{.preserve}{not used}
}
\value{
a filtered \link{memoMDB} object
}
\description{
Filter a \link{metaMDB} object
}
