% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{filter_ch_matrix}
\alias{filter_ch_matrix}
\title{Filter a matrix stored in a chMDB}
\usage{
filter_ch_matrix(x, tableName, ...)
}
\arguments{
\item{x}{a \link{chMDB} object}

\item{tableName}{a character vector of length 1 corresponding to the name of
the table to filter (must be a matrix)}

\item{...}{character vectors with the row names and/or columns names to
select. The names of the parameters must correspond to the name of the
column and of the row fields (the matrix cannot be filtered from values).}
}
\value{
A sub-matrix of tableName in x
}
\description{
Filter a matrix stored in a chMDB
}
\examples{
\dontrun{
## Return the matrix of expression values focused on the selected genes
filter_ch_matrix(x=db, "Expression_value", gene=c("SNCA", "MAPT"))
}

}
