% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoMDB.R
\name{filter_with_tables.memoMDB}
\alias{filter_with_tables.memoMDB}
\title{Filter \link{memoMDB} object according to provided tables}
\usage{
\method{filter_with_tables}{memoMDB}(x, tables, checkTables = TRUE)
}
\arguments{
\item{x}{a \link{memoMDB} object}

\item{tables}{a named list of tibbles to filter with. The names should
correspond to the table names in x and the tibbles should fit the
data model.}

\item{checkTables}{if TRUE, the tables are confronted to their model
in the data model of x.}
}
\value{
a \link{memoMDB} object
}
\description{
Filter \link{memoMDB} object according to provided tables
}
