% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memoMDB.R,
%   R/metaMDB.R
\name{filter_with_tables.fileMDB}
\alias{filter_with_tables.fileMDB}
\alias{filter_with_tables}
\alias{filter_with_tables.memoMDB}
\alias{filter_with_tables.metaMDB}
\title{Filter \link{fileMDB} object according to provided tables}
\usage{
\method{filter_with_tables}{fileMDB}(x, tables, checkTables = TRUE)

filter_with_tables(x, tables, checkTables = TRUE)

\method{filter_with_tables}{memoMDB}(x, tables, checkTables = TRUE)

\method{filter_with_tables}{metaMDB}(x, tables, checkTables = TRUE)
}
\arguments{
\item{x}{a \link{metaMDB} object}

\item{tables}{a named list of tibbles to filter with. The names should
correspond to the table names in x and the tibbles should fit the
data model.}

\item{checkTables}{if TRUE, the tables are confronted to their model
in the data model of x.}
}
\value{
a \link{memoMDB} object

a \link{memoMDB} object

a \link{memoMDB} object

a \link{metaMDB} object
}
\description{
Filter \link{fileMDB} object according to provided tables

Filter an MDB object according to provided tables

Filter \link{memoMDB} object according to provided tables

Filter \link{metaMDB} object according to provided tables
}
