% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{format.chMDB}
\alias{format.chMDB}
\title{Format a \link{chMDB} object for printing}
\usage{
\method{format}{chMDB}(x)
}
\arguments{
\item{x}{a \link{chMDB} object}
}
\value{
A single character
}
\description{
Format a \link{chMDB} object for printing
}
