% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{format.chTKCat}
\alias{format.chTKCat}
\title{Format an \link{chTKCat} object for printing}
\usage{
\method{format}{chTKCat}(x)
}
\arguments{
\item{x}{an \link{chTKCat} object}
}
\value{
A single character
}
\description{
Format an \link{chTKCat} object for printing
}
