% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{format.internalMDB}
\alias{format.internalMDB}
\title{Format an \link{internalMDB} object for printing}
\usage{
\method{format}{internalMDB}(x)
}
\arguments{
\item{x}{an \link{internalMDB} object}
}
\value{
A single character
}
\description{
Format an \link{internalMDB} object for printing
}
