% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMDB.R
\name{full_join_MDBs}
\alias{full_join_MDBs}
\title{Join 2 MDBs}
\usage{
full_join_MDBs(
  x,
  y,
  by = get_shared_collections(x, y),
  dbInfo = list(name = paste(db_info(x)$name, db_info(y)$name, sep = "_")),
  dmAutoLayout = TRUE,
  ...
)
}
\arguments{
\item{x}{an MDB}

\item{y}{an MDB}

\item{by}{a tibble as returned by the \code{\link[=get_shared_collections]{get_shared_collections()}} function
which indicates which collection members should be joined through
a relational table. If the collection is \code{NA}, the relational table is built
by merging identical columns in table.x and table.y. If the collection
is provided, the relational table is build using
the \code{\link[=map_collection_members]{map_collection_members()}} function.}

\item{dbInfo}{a list with DB information:
\strong{"name"} (only mandatory field), "title", "description", "url",
"version", "maintainer".}

\item{dmAutoLayout}{if TRUE (default) the layout of the joined data model
is automatically adjusted.}

\item{...}{further parameters for the \code{\link[=map_collection_members]{map_collection_members()}} function.}
}
\value{
a \link{metaMDB} object
}
\description{
Join 2 MDBs
}
