% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{get_chMDB}
\alias{get_chMDB}
\title{Get a \link{chMDB} from a \link{chTKCat}}
\usage{
get_chMDB(tkcon, dbName, n_max = 10)
}
\arguments{
\item{tkcon}{a \link{chTKCat} object}

\item{dbName}{the name of the database}

\item{n_max}{maximum number of records to read
for checks purpose (default: 10). See also \code{\link[ReDaMoR:confront_data]{ReDaMoR::confront_data()}}.}
}
\description{
Get a \link{chMDB} from a \link{chTKCat}
}
