% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{get_chMDB_metadata}
\alias{get_chMDB_metadata}
\title{Get the metadata of an MDB from a \link{chTKCat} connection}
\usage{
get_chMDB_metadata(x, dbName, timestamp = NA)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{dbName}{the name of the MDB}

\item{timestamp}{the timestamp of the instance to get.
Default=NA: get the current version.}
}
\value{
A list with the following elements:
\itemize{
\item dbInfo: General information regarding the MDB
\item dataModel: The data model
\item collectionMembers: Members of different collections
\item access: type of access to the MDB
}
}
\description{
Get the metadata of an MDB from a \link{chTKCat} connection
}
\seealso{
\link{get_MDB}
}
