% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-collections.R
\name{get_collection_mapper}
\alias{get_collection_mapper}
\title{Get the default mapper function for a collection}
\usage{
get_collection_mapper(collection)
}
\arguments{
\item{collection}{the name of the targeted collection
(it should belong to local collections: see \code{\link[=list_local_collections]{list_local_collections()}}).}
}
\value{
A function to map collection members.
}
\description{
Get the default mapper function for a collection
}
