% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDB.R
\name{get_confrontation_report}
\alias{get_confrontation_report}
\title{Get the last generated MDB confrontation report}
\usage{
get_confrontation_report()
}
\value{
A confrontation report generated by \code{\link[ReDaMoR:confront_data]{ReDaMoR::confront_data()}}
}
\description{
Get the last generated MDB confrontation report
}
