% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R, R/chMDB.R, R/chTKCat.R,
%   R/generics.R, R/metaMDB.R
\name{get_hosts.DBIConnection}
\alias{get_hosts.DBIConnection}
\alias{get_hosts.chMDB}
\alias{get_hosts.chTKCat}
\alias{get_hosts}
\alias{get_hosts.metaMDB}
\title{Get database hosts}
\usage{
\method{get_hosts}{DBIConnection}(x, ...)

\method{get_hosts}{chMDB}(x, ...)

\method{get_hosts}{chTKCat}(x, ...)

get_hosts(x, ...)

\method{get_hosts}{metaMDB}(x, ...)
}
\arguments{
\item{x}{an object with database connection(s)}

\item{...}{additional parameters for methods.}
}
\value{
A character vector with hosts information (generaly 1)
in the following shape: "host:port"
}
\description{
Get database hosts
}
