% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM-specifications.R
\name{get_km_spec}
\alias{get_km_spec}
\title{Get KM specifications from an \link{MDB} object}
\usage{
get_km_spec(x, kmr)
}
\arguments{
\item{x}{an \link{MDB} object with specification tables}

\item{kmr}{an \link{MDB} object with KM requirements}
}
\value{
An \link{MDB} object with kmr specification tables from x
}
\description{
Get KM specifications from an \link{MDB} object
}
