% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{get_KMR}
\alias{get_KMR}
\title{Get a \link{KMR} object from a \link{TKCat} or a \link{chTKCat} object}
\usage{
get_KMR(...)
}
\arguments{
\item{...}{parameters for the \code{\link[=get_MDB]{get_MDB()}} function}
}
\value{
A \link{KMR} object
}
\description{
Get a \link{KMR} object from a \link{TKCat} or a \link{chTKCat} object
}
