% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-collections.R
\name{get_local_collection}
\alias{get_local_collection}
\title{Get the json definition of a local collection of concepts}
\usage{
get_local_collection(title)
}
\arguments{
\item{title}{the title of the collection to get}
}
\value{
The definition of the collection as a JSON string.
}
\description{
Get the json definition of a local collection of concepts
}
