% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKCat.R, R/chMDB.R, R/generics.R
\name{get_MDB.TKCat}
\alias{get_MDB.TKCat}
\alias{get_MDB.chTKCat}
\alias{get_MDB}
\title{Get an \link{MDB} object from a \link{TKCat} related object}
\usage{
\method{get_MDB}{TKCat}(x, dbName, ...)

\method{get_MDB}{chTKCat}(x, dbName, timestamp = NA, check = TRUE, n_max = 10, ...)

get_MDB(x, dbName, ...)
}
\arguments{
\item{x}{a \link{TKCat} related object (e.g. \link{chTKCat})}

\item{dbName}{the name of the database}

\item{...}{method specific parameters}

\item{timestamp}{the timestamp of the instance to get.
Default=NA: get the current version.}

\item{check}{logical: if TRUE (default) the data are confronted to the
data model}

\item{n_max}{maximum number of records to read
for checks purpose (default: 10). See also \code{\link[ReDaMoR:confront_data]{ReDaMoR::confront_data()}}.}
}
\value{
An \link{MDB} object
}
\description{
Get an \link{MDB} object from a \link{TKCat} related object
}
\seealso{
\link{get_confrontation_report}, \link[ReDaMoR:format_confrontation_report]{ReDaMoR::format_confrontation_report}
and \link[ReDaMoR:format_confrontation_report_md]{ReDaMoR::format_confrontation_report_md} for getting and formatting
the report confronting the data to the model.
}
