% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge-access.R
\name{get_pok}
\alias{get_pok}
\title{Get a piece of knowledge from a \link{chTKCat} connection}
\usage{
get_pok(x, mdb_name, corpus_name)
}
\arguments{
\item{x}{a \link{chTKCat} or a \link[=chMDB]{KMR} object}

\item{mdb_name}{the piece of knowledge to get}

\item{corpus_name}{if x is a \link{chTKCat} object,
the corpus to which the piece belongs. Otherwise this parameter is ignored.}
}
\description{
Get a piece of knowledge from a \link{chTKCat} connection
}
