% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/chTKCat.R, R/generics.R
\name{get_query.chMDB}
\alias{get_query.chMDB}
\alias{get_query.chTKCat}
\alias{get_query}
\title{Get SQL query}
\usage{
\method{get_query}{chMDB}(x, query, ...)

\method{get_query}{chTKCat}(x, query, ...)

get_query(x, query, ...)
}
\arguments{
\item{x}{an object with a database connection}

\item{query}{the SQL query}

\item{...}{method specific parameters}
}
\value{
A tibble with query results
}
\description{
Get SQL query
}
