% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_kmspec.R
\name{get_R_helpers}
\alias{get_R_helpers}
\title{Get helpers from an \link{MDB} object}
\usage{
get_R_helpers(x, kmr, name = "R-Helpers")
}
\arguments{
\item{x}{the MDB object}

\item{kmr}{an \link{MDB} object with KM requirements}

\item{name}{the name of the helper set (default: "R-Helpers")}
}
\value{
A list of functions with their environment.
See \code{\link[=parse_R_helpers]{parse_R_helpers()}}.
}
\description{
Get helpers from an \link{MDB} object
}
