% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMDB.R
\name{get_shared_collections}
\alias{get_shared_collections}
\title{Get collections shared by 2 objects and return member combinations}
\usage{
get_shared_collections(x, y)
}
\arguments{
\item{x}{an MDB object}

\item{y}{an MDB object}
}
\value{
A tibble with the following fields:
\itemize{
\item \strong{collection} the name of the collection
\item \strong{mid.x} the collection member identifier in x
\item \strong{table.x} the table of the collection member in x
\item \strong{mid.y} the collection member identifier in y
\item \strong{table.y} the table of the collection member in y
}
}
\description{
Get collections shared by 2 objects and return member combinations
}
