% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM-specifications.R
\name{has_km_spec}
\alias{has_km_spec}
\title{Check if KM specifications are available in an \link{MDB} object}
\usage{
has_km_spec(x, kmr)
}
\arguments{
\item{x}{an \link{MDB} object with specification tables}

\item{kmr}{an \link{MDB} object with KM requirements}
}
\value{
A logical: TRUE if the MDB has KM specifications
}
\description{
Check if KM specifications are available in an \link{MDB} object
}
