% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/fileMDB.R, R/generics.R,
%   R/memoMDB.R, R/metaMDB.R
\name{heads.chMDB}
\alias{heads.chMDB}
\alias{heads.fileMDB}
\alias{heads}
\alias{heads.memoMDB}
\alias{heads.metaMDB}
\title{Get the first records of each object data tables}
\usage{
\method{heads}{chMDB}(x, ..., n = 6L)

\method{heads}{fileMDB}(x, ..., n = 6L)

heads(x, ..., n = 6L)

\method{heads}{memoMDB}(x, ..., n = 6L)

\method{heads}{metaMDB}(x, ..., n = 6L)
}
\arguments{
\item{x}{an object with embedded data tables}

\item{...}{the name of the tables to get (default: all of them)}

\item{n}{maximum number of records to return (default: 6)}
}
\value{
A list of \link[dplyr:reexports]{dplyr::tibble} and \link{matrix}
}
\description{
Get the first records of each object data tables
}
