% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-collections.R
\name{import_collection_mapper}
\alias{import_collection_mapper}
\title{Import a  function to map collection members}
\usage{
import_collection_mapper(collection, fun)
}
\arguments{
\item{collection}{the name of the targeted collection
(it should belong to local collections: see \code{\link[=list_local_collections]{list_local_collections()}}).}

\item{fun}{a function which takes 2 data.frames (x an y) with
fields described in the collection definition and map the different elements.}
}
\description{
Import a  function to map collection members
}
