% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-collections.R
\name{import_local_collection}
\alias{import_local_collection}
\title{Import a  the definition of a collection of concepts in the local environment}
\usage{
import_local_collection(txt, overwrite = FALSE)
}
\arguments{
\item{txt}{a JSON string or file}

\item{overwrite}{a single logical. If TRUE the collection is overwritten
if it already exists (default: FALSE)}
}
\value{
No return value, called for side effects. The collection will be
available and operations will be possible on its members.
}
\description{
Import a  the definition of a collection of concepts in the local environment
}
