% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importLocalCollection.R
\name{importLocalCollection}
\alias{importLocalCollection}
\title{Import a collection definition in the local environment}
\usage{
importLocalCollection(f)
}
\arguments{
\item{f}{the JSON file to import}
}
\description{
Import a collection definition in the local environment
}
