% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{internalMDB}
\alias{internalMDB}
\title{Create an internal database: a set of tables with a data model}
\usage{
internalMDB(
  dataModel,
  dbTables,
  dbInfo,
  colMembers = NULL,
  checks = c("unique", "not nullable", "foreign keys"),
  verbose = FALSE
)
}
\arguments{
\item{dataModel}{a RelDataModel object}

\item{dbInfo}{a list of information values regarding the DB}

\item{colMembers}{the collection members of the MDB. If NULL (default),
collection members are not defined.}

\item{checks}{a character vector with the name of optional checks to be
done (all of them c("unique", "not nullable", "foreign keys"))}

\item{verbose}{if TRUE display the data confrontation report}

\item{dbTable}{a list of tables}
}
\value{
An internalMDB object
}
\description{
Create an internal database: a set of tables with a data model
}
