% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{is.chTKCat}
\alias{is.chTKCat}
\title{Check the object is  an \link{chTKCat} object}
\usage{
is.chTKCat(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is an chTKCat object
}
\description{
Check the object is  an \link{chTKCat} object
}
