% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{is.fileMDB}
\alias{is.fileMDB}
\title{Check the object is  a \link{fileMDB} object}
\usage{
is.fileMDB(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is an \link{fileMDB} object
}
\description{
Check the object is  a \link{fileMDB} object
}
