% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{is.internalMDB}
\alias{is.internalMDB}
\title{Check the object is  an \link{internalMDB} object}
\usage{
is.internalMDB(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is an internalMDB object
}
\description{
Check the object is  an \link{internalMDB} object
}
