% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memeMDB.R
\name{is.memeMDB}
\alias{is.memeMDB}
\title{Check the object is  a \link{memeMDB} object}
\usage{
is.memeMDB(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is an \link{memeMDB} object
}
\description{
Check the object is  a \link{memeMDB} object
}
