% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoMDB.R
\name{is.memoMDB}
\alias{is.memoMDB}
\title{Check if the object is  a \link{memoMDB} object}
\usage{
is.memoMDB(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is an \link{memoMDB} object
}
\description{
Check if the object is  a \link{memoMDB} object
}
