% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMDB.R
\name{is.metaMDB}
\alias{is.metaMDB}
\title{Check the object is  a \link{metaMDB} object}
\usage{
is.metaMDB(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is an \link{metaMDB} object
}
\description{
Check the object is  a \link{metaMDB} object
}
