% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POK.R
\name{is.POK}
\alias{is.POK}
\title{Check if the object is a \link{POK} object}
\usage{
is.POK(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is a \link{POK} object
}
\description{
Check if the object is a \link{POK} object
}
