% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKCat.R
\name{is.TKCat}
\alias{is.TKCat}
\title{Check the object is  a \link{TKCat} object}
\usage{
is.TKCat(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is a \link{TKCat} object
}
\description{
Check the object is  a \link{TKCat} object
}
