% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{is_chMDB_public}
\alias{is_chMDB_public}
\title{Is a chMDB public}
\usage{
is_chMDB_public(x, mdb)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{mdb}{name of the modeled database}
}
\value{
A logical indicating if the chMDB is public or not.
}
\description{
Is a chMDB public
}
