% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{is_current_chMDB}
\alias{is_current_chMDB}
\title{Check if the \link{chMDB} object refers to the current instance of the MDB}
\usage{
is_current_chMDB(x)
}
\arguments{
\item{x}{a \link{chMDB} object}
}
\value{
A single logical:
TRUE if x refers to the current instance of the MDB.
}
\description{
Check if the \link{chMDB} object refers to the current instance of the MDB
}
