% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_kmr.R
\name{is_KMR}
\alias{is_KMR}
\title{Check if an object represents Knowledge Management Requirements (KMR)}
\usage{
is_KMR(x)
}
\arguments{
\item{x}{an object}
}
\value{
TRUE if x is a KMR, FALSE if not
}
\description{
Check if an object represents Knowledge Management Requirements (KMR)
}
