% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{length.chMDB}
\alias{length.chMDB}
\title{Get the number of tables in a \link{chMDB} object}
\usage{
\method{length}{chMDB}(x)
}
\arguments{
\item{x}{a \link{chMDB} object}
}
\value{
A single integer value
}
\description{
Get the number of tables in a \link{chMDB} object
}
