% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{length.internalMDB}
\alias{length.internalMDB}
\title{Get the number of tables in an \link{internalMDB} object}
\usage{
\method{length}{internalMDB}(x)
}
\arguments{
\item{x}{an \link{internalMDB} object}
}
\value{
A single integer value
}
\description{
Get the number of tables in an \link{internalMDB} object
}
