% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDB.R
\name{length.MDB}
\alias{length.MDB}
\title{Get the number of tables in an MDB object}
\usage{
\method{length}{MDB}(x)
}
\arguments{
\item{x}{an MDB object}
}
\value{
A single integer value
}
\description{
Get the number of tables in an MDB object
}
