% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDB.R
\name{lengths.MDB}
\alias{lengths.MDB}
\title{Get the number of fields in each table of an MDB object}
\usage{
\method{lengths}{MDB}(x, use.names = TRUE)
}
\arguments{
\item{x}{an MDB object}

\item{use.names}{return the names of the tables}
}
\value{
A vector of integers
}
\description{
Get the number of fields in each table of an MDB object
}
