% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{list_chMDB_timestamps}
\alias{list_chMDB_timestamps}
\title{List instance timestamps of an MDB in \link{chTKCat}}
\usage{
list_chMDB_timestamps(x, name)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the database}
}
\value{
A tibble with the instance of each table at each timestamp.
The "current" attribute indicate the current timestamp instance.
If there is no recorded timestamp, the function returns NULL.
}
\description{
List instance timestamps of an MDB in \link{chTKCat}
}
