% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{list_chMDB_users}
\alias{list_chMDB_users}
\title{List users of an MDB of a \link{chTKCat} object}
\usage{
list_chMDB_users(x, mdb)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{mdb}{name of the modeled database}
}
\value{
A tibble with 3 columns:
\itemize{
\item user: the user login
\item mdb: the name of the modeled database
\item admin: if the user is an admin of the MDB
}
}
\description{
List users of an MDB of a \link{chTKCat} object
}
