% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{list_chMDBs}
\alias{list_chMDBs}
\title{List available databases in a \link{chTKCat}}
\usage{
list_chMDBs(x, withInfo = TRUE)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{withInfo}{if TRUE (default), it returns only initialized MDBs along
with general information. If FALSE, it returns all the available databases
in the system (excepted those reserved by the system).}
}
\description{
List available databases in a \link{chTKCat}
}
