% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{list_chTKCat_collections}
\alias{list_chTKCat_collections}
\title{List collections available in a \link{chTKCat}}
\usage{
list_chTKCat_collections(x)
}
\arguments{
\item{x}{a \link{chTKCat} object}
}
\description{
List collections available in a \link{chTKCat}
}
