% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{list_chTKCat_users}
\alias{list_chTKCat_users}
\title{List \link{chTKCat} user}
\usage{
list_chTKCat_users(x)
}
\arguments{
\item{x}{a \link{chTKCat} object}
}
\value{
A tibble with 3 columns:
\itemize{
\item login: user login
\item contact: user contact information
\item admin: if the user is an admin of the chTKCat object
}
}
\description{
List \link{chTKCat} user
}
