% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{list_feature_properties}
\alias{list_feature_properties}
\title{List properties of a feature}
\usage{
list_feature_properties(kmr, feature)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{feature}{the name of the feature}
}
\value{
A \link[dplyr:reexports]{dplyr::tibble} with the description of feature properties
}
\description{
List properties of a feature
}
