% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-collections.R
\name{list_local_collections}
\alias{list_local_collections}
\title{List local collections of concepts}
\usage{
list_local_collections(withJson = FALSE)
}
\arguments{
\item{withJson}{if TRUE, returns the json strings of the collection
(default: FALSE)}
}
\value{
A tibble with the title, the description and optionally the json
definition of the collections
}
\description{
List local collections of concepts
}
