% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKCat.R, R/chTKCat.R, R/generics.R
\name{list_MDBs.TKCat}
\alias{list_MDBs.TKCat}
\alias{list_MDBs.chTKCat}
\alias{list_MDBs}
\title{List available \link{MDB}}
\usage{
\method{list_MDBs}{TKCat}(x, withInfo = TRUE)

\method{list_MDBs}{chTKCat}(x, withInfo = TRUE)

list_MDBs(x, withInfo = TRUE)
}
\arguments{
\item{x}{a \link{TKCat} related object (e.g. \link{chTKCat})}

\item{withInfo}{if TRUE (default), the function returns a table with
\link{db_info}. If FALSE, it returns only \link{MDB} names.}
}
\value{
A tibble with information about the \link{MDB} available in
a \link{TKCat} related object.
}
\description{
List available \link{MDB}
}
