% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{list_measurement_units}
\alias{list_measurement_units}
\title{List possible units for a type of measurement}
\usage{
list_measurement_units(kmr, measurement)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{measurement}{the type of measurement}
}
\value{
A \link{tibble} with the description of supported units
}
\description{
List possible units for a type of measurement
}
