% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{list_measurements}
\alias{list_measurements}
\title{List supported types of measurement}
\usage{
list_measurements(kmr)
}
\arguments{
\item{kmr}{a \link{KMR} object}
}
\value{
A vector of character with measurement names
}
\description{
List supported types of measurement
}
