% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POK.R
\name{list_POKs}
\alias{list_POKs}
\title{List available \link{POK}}
\usage{
list_POKs(x, kmr)
}
\arguments{
\item{x}{a \link{chTKCat} or \link{TKCat} object}

\item{kmr}{\link{KMR} object with KM requirements or its name in tkcat}
}
\value{
The names of available POKs in x with kmr requirements.
}
\description{
List available \link{POK}
}
