% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{list_property_values}
\alias{list_property_values}
\title{List of supported values for an integer or a character property}
\usage{
list_property_values(kmr, feature, property)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{feature}{the name of the feature}

\item{property}{the name of the property}
}
\value{
A \link[dplyr:reexports]{dplyr::tibble} with value and their description (if available)
}
\description{
List of supported values for an integer or a character property
}
