% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{list_table_features}
\alias{list_table_features}
\title{List the features provided by a set of tables}
\usage{
list_table_features(kmr, tables = NULL)
}
\arguments{
\item{kmr}{a \link{KMR} object}

\item{tables}{the name of the tables. If NULL (default), all the features
are listed.}
}
\value{
A \link{tibble} with feature description and properties
}
\description{
List the features provided by a set of tables
}
