% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{list_table_types}
\alias{list_table_types}
\title{List types of tables defined in a \link{KMR} object}
\usage{
list_table_types(kmr)
}
\arguments{
\item{kmr}{a \link{KMR} object}
}
\value{
A \link{tibble} with the names of table types, their descriptions and
the related collections
}
\description{
List types of tables defined in a \link{KMR} object
}
