% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R
\name{list_tables}
\alias{list_tables}
\title{List tables in a clickhouse database}
\usage{
list_tables(con, dbNames = NULL)
}
\arguments{
\item{con}{the clickhouse connection}

\item{dbNames}{the name of databases to focus on (default NULL ==> all)}
}
\value{
A tibble with the following columns:
\itemize{
\item \strong{database}: the name of the database
\item \strong{name}: the name of the table
\item \strong{total_rows}: the number of rows in the table
\item \strong{total_bytes}: the size of the table
}
}
\description{
List tables in a clickhouse database
}
