% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{listChTKCatCollections}
\alias{listChTKCatCollections}
\title{List collections available in a \link{chTKCat}}
\usage{
listChTKCatCollections(tkcon)
}
\arguments{
\item{tkcon}{a \link{chTKCat} object}
}
\description{
List collections available in a \link{chTKCat}
}
