% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{manage_chTKCat_users}
\alias{manage_chTKCat_users}
\title{Manage user information in a shiny interface}
\usage{
manage_chTKCat_users(x, pwdFile = NULL)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{pwdFile}{a local file in which the password for x can be found.
If NULL (default), the connection is shared by all sessions and can
be disabled at some point.}
}
\description{
Manage user information in a shiny interface
}
