% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{materialize_chMDB_projections}
\alias{materialize_chMDB_projections}
\title{Materialize in ClickHouse any projection documented in the data model}
\usage{
materialize_chMDB_projections(x)
}
\arguments{
\item{x}{a \link{chMDB} object}
}
\value{
Nothing: materialize projections in ClickHouse.
}
\description{
Materialize in ClickHouse any projection documented in the data model
}
