% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMDB.R
\name{MDBs}
\alias{MDBs}
\title{Get a list of MDB from \link{metaMDB} object}
\usage{
MDBs(x)
}
\arguments{
\item{x}{a \link{metaMDB} object}
}
\value{
A list of MDB objects
}
\description{
Get a list of MDB from \link{metaMDB} object
}
