% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memeMDB.R
\name{memeMDB}
\alias{memeMDB}
\title{Create an MDB (Modeled DataBase) in memory: memeMDB}
\usage{
memeMDB(
  dataTables,
  dataModel,
  dbInfo,
  collectionMembers = NULL,
  checks = c("unique", "not nullable", "foreign keys"),
  verbose = FALSE
)
}
\arguments{
\item{dataTables}{a list of tibbles}

\item{dataModel}{a \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object}

\item{dbInfo}{a list with DB information:
"name", "title", "description", "url",
"version", "maintainer".}

\item{collectionMembers}{the members of collections as provided to the
\link{collection_members<-} function (default: NULL ==> no member).}

\item{checks}{a character vector with the name of optional checks to be
done (all of them c("unique", "not nullable", "foreign keys"))}

\item{verbose}{if TRUE display the data confrontation report
(default: FALSE)}
}
\value{
A memeMDB object
}
\description{
Create an MDB (Modeled DataBase) in memory: memeMDB
}
