% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoMDB.R
\name{memoMDB}
\alias{memoMDB}
\title{Create an MDB (Modeled DataBase) in memory: memoMDB}
\usage{
memoMDB(
  dataTables,
  dataModel,
  dbInfo,
  collectionMembers = NULL,
  checks = c("unique", "not nullable", "foreign keys"),
  verbose = FALSE
)
}
\arguments{
\item{dataTables}{a list of tibbles}

\item{dataModel}{a \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object}

\item{dbInfo}{a list with DB information:
\strong{"name"} (only mandatory field), "title", "description", "url",
"version", "maintainer".}

\item{collectionMembers}{the members of collections as provided to the
\link{collection_members<-} function (default: NULL ==> no member).}

\item{checks}{a character vector with the name of optional checks to be
done (all of them c("unique", "not nullable", "foreign keys"))}

\item{verbose}{if TRUE display the data confrontation report
(default: FALSE)}
}
\value{
A memoMDB object
}
\description{
Create an MDB (Modeled DataBase) in memory: memoMDB
}
\examples{
library(devTKCat)

hpo <- read_fileMDB(
   path=system.file("examples/HPO-subset", package="ReDaMoR"),
   dataModel=system.file("examples/HPO-model.json", package="ReDaMoR"),
   dbInfo=list(
      "name"="HPO",
      "title"="Data extracted from the HPO database",
      "description"=paste(
         "This is a very small subset of the HPO!",
         "Visit the reference URL for more information"
      ),
      "url"="http://human-phenotype-ontology.github.io/"
   )
) \%>\% 
   as_memoMDB()
count_records(hpo)

hpoSlice <- slice(hpo, HPO_diseases=1:10)
count_records(hpoSlice)

if(requireNamespace("stringr", quietly = TRUE)){
   epilHP <- filter(
      hpo,
      HPO_diseases=stringr::str_detect(
         label, stringr::regex("epilepsy", ignore_case=TRUE)
      )
   )
   count_records(epilHP)
}
}
