% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R
\name{mergeTree_from_RelTableModel}
\alias{mergeTree_from_RelTableModel}
\title{Create a ClickHouse MergeTree table from a \link[ReDaMoR:RelTableModel]{ReDaMoR::RelTableModel}}
\usage{
mergeTree_from_RelTableModel(con, dbName, tm)
}
\arguments{
\item{con}{the clickhouse connection}

\item{dbName}{the name of the database in which the table should be written}

\item{tm}{a \link[ReDaMoR:RelTableModel]{ReDaMoR::RelTableModel} object}
}
\value{
No return value, called for side effects
}
\description{
Create a ClickHouse MergeTree table from a \link[ReDaMoR:RelTableModel]{ReDaMoR::RelTableModel}
}
