% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R
\name{mergeTrees_from_RelDataModel}
\alias{mergeTrees_from_RelDataModel}
\title{Create ClickHouse MergeTree tables from a \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel}}
\usage{
mergeTrees_from_RelDataModel(con, dbName, dbm)
}
\arguments{
\item{con}{the clickhouse connection}

\item{dbName}{the name of the database in which the tables should be written}

\item{dbm}{a \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object}
}
\value{
No return value, called for side effects
}
\description{
Create ClickHouse MergeTree tables from a \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel}
}
