% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaMDB.R
\name{metaMDB}
\alias{metaMDB}
\title{Create a metaMDB object}
\usage{
metaMDB(MDBs, relationalTables, dataModel, dbInfo)
}
\arguments{
\item{MDBs}{a list of MDB objects}

\item{relationalTables}{a list of tibbles corresponding to the relational
tables between the different MDBs}

\item{dataModel}{a \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object gathering all the
data model of all the MDBs plus the relational tables}

\item{dbInfo}{a list with DB information:
\strong{"name"} (only mandatory field), "title", "description", "url",
"version", "maintainer".}
}
\value{
A metaMDB object
}
\description{
A metaMDB object is an MDB gathering several other MDBs glued by
relational tables.
}
