% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-management.R
\name{parse_R_helpers}
\alias{parse_R_helpers}
\title{Parse source code to get R helpers}
\usage{
parse_R_helpers(code)
}
\arguments{
\item{code}{the source code as a character vector}
}
\value{
A list of functions from \code{code} plus an "help" function used to get
function documentation.
}
\description{
Parse source code to get R helpers
}
\details{
Functions in code must be documented
with \link[roxygen2:roxygen2-package]{roxygen2::roxygen2-package}
tags and only functions with the '@export' tag are returned.
}
\examples{
{

code <- "
   a <- 2
   #' Set the 'a' value to use in the add_a function
   #' 
   #' @param v a numeric value
   #' 
   #' @return Does not return anything
   #' 
   #' @export
   set_a <- function(v) a <<- v
   
   #' Add a 'a' value defined separately
   #' 
   #' @param x a numeric value
   #' 
   #' @return x + a
   #' 
   #' @export
   add_a <- function(x) x + a
"
helpers <- parse_R_helpers(code)
helpers$help()
helpers$help("add_a")
helpers$add_a(3.5)
helpers$set_a(4)
helpers$add_a(3.5)

}

}
