% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{plot.chMDB}
\alias{plot.chMDB}
\title{Plot the underlying \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel}}
\usage{
\method{plot}{chMDB}(x)
}
\description{
Plot the underlying \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel}
}
\seealso{
\link[ReDaMoR:plot.RelDataModel]{ReDaMoR::plot.RelDataModel}
}
