% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{plot.internalMDB}
\alias{plot.internalMDB}
\title{Plot the underlying \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel}}
\usage{
\method{plot}{internalMDB}(x)
}
\description{
Plot the underlying \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel}
}
\seealso{
\link[ReDaMoR:plot.RelDataModel]{ReDaMoR::plot.RelDataModel}
}
