% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_collection_members.R
\name{read_collection_members}
\alias{read_collection_members}
\title{Read a collection member JSON file}
\usage{
read_collection_members(f)
}
\arguments{
\item{f}{the JSON file to read}
}
\value{
A tibble with the description of the collection members of a
resource
}
\description{
Read a collection member JSON file
}
