% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMR.R
\name{read_KMR}
\alias{read_KMR}
\title{Read \link{KMR} from a path}
\usage{
read_KMR(...)
}
\arguments{
\item{...}{parameters for \code{\link[=read_fileMDB]{read_fileMDB()}}}
}
\value{
A \link{KMR} object
}
\description{
Read \link{KMR} from a path
}
