% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCollectionMembers.R
\name{readCollectionMembers}
\alias{readCollectionMembers}
\title{Read a collection member JSON file}
\usage{
readCollectionMembers(f)
}
\arguments{
\item{f}{the JSON file to read}
}
\value{
A tibble with the descripition of the collection members of a
resource
}
\description{
Read a collection member JSON file
}
