% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInternalMDB.R
\name{readInternalMDB}
\alias{readInternalMDB}
\title{Read files and create an internal database with a documented data model}
\usage{
readInternalMDB(
  dataModel,
  descriptionFile,
  directory,
  verbose = FALSE,
  cr_variable_name = NULL,
  ext = "txt",
  delim = "\\t",
  quoted_na = FALSE,
  ...
)
}
\arguments{
\item{dataModel}{a \link{RelDataModel} object}

\item{descriptionFile}{a json file with DB information}

\item{directory}{a path to the directory where to read the data}

\item{verbose}{if TRUE display the data confrontation report}

\item{cr_variable_name}{a character with a single value indicating the name
of the variable in the .GlobalEnv environment where to put the data
confrontation report. If it's not a character with a single
value (default: NULL), then the confrontation report is not exported.}

\item{ext}{the file extension to consider (default: "ext"),}

\item{delim}{delimiter (default: '\\\\t')}

\item{quoted_na}{Should missing values inside quotes be treated
as missing values or as strings or strings (the default).
Be aware that the default value here is different than the one for the
original \code{\link[readr:read_delim]{readr::read_delim()}} function.}

\item{...}{additional  parameters for the \code{\link[readr:read_delim]{readr::read_delim()}} function}
}
\value{
An \link{internalMDB} object
}
\description{
Read files and create an internal database with a documented data model
}
