% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{reconnect_chTKCat}
\alias{reconnect_chTKCat}
\title{Reconnect to a \link{chTKCat} instance}
\usage{
reconnect_chTKCat(x, ...)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{...}{further arguments for \link{ch_reconnect}:
\itemize{
\item \strong{user}: user name. If not provided, it's taken from x
\item \strong{password}: user password. If not provided, first the function
tries to connect without any password. If it fails, the function asks the
user to provide a password.
\item \strong{ntries}: the number of times the user can enter a wrong password
}}
}
\description{
Reconnect to a \link{chTKCat} instance
}
