% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{remove_chMDB_user}
\alias{remove_chMDB_user}
\title{Drop a user of an MDB of a \link{chTKCat} object}
\usage{
remove_chMDB_user(x, mdb, login)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{mdb}{name of the modeled database}

\item{login}{login of the user to drop}
}
\value{
No return value, called for side effects
}
\description{
Drop a user of an MDB of a \link{chTKCat} object
}
