% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{remove_chTKCat_collection}
\alias{remove_chTKCat_collection}
\title{Remove a collection from a \link{chTKCat} database}
\usage{
remove_chTKCat_collection(tkcon, name)
}
\arguments{
\item{tkcon}{a \link{chTKCat} object}

\item{name}{the name of the collection to remove}
}
\description{
Remove a collection from a \link{chTKCat} database
}
