% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{rename.chMDB}
\alias{rename.chMDB}
\title{Rename tables of a \link{chMDB} object}
\usage{
\method{rename}{chMDB}(.data, ...)
}
\arguments{
\item{.data}{a \link{chMDB} object}

\item{...}{Use new_name = old_name to rename selected tables}
}
\description{
Rename tables of a \link{chMDB} object
}
