% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{rename.fileMDB}
\alias{rename.fileMDB}
\title{Rename tables of a \link{fileMDB} object}
\usage{
\method{rename}{fileMDB}(.data, ...)
}
\arguments{
\item{.data}{a \link{fileMDB} object}

\item{...}{Use new_name = old_name to rename selected tables}
}
\description{
Rename tables of a \link{fileMDB} object
}
