% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoMDB.R
\name{rename.memoMDB}
\alias{rename.memoMDB}
\title{Rename tables of a \link{memoMDB} object}
\usage{
\method{rename}{memoMDB}(.data, ...)
}
\arguments{
\item{.data}{a \link{memoMDB} object}

\item{...}{Use new_name = old_name to rename selected tables}
}
\description{
Rename tables of a \link{memoMDB} object
}
