% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKCat.R
\name{rename.TKCat}
\alias{rename.TKCat}
\title{Rename a \link{TKCat} object}
\usage{
\method{rename}{TKCat}(.data, ...)
}
\arguments{
\item{.data}{a \link{TKCat} object}

\item{...}{Use new_name = old_name to rename selected \link{MDB}}
}
\description{
Rename a \link{TKCat} object
}
