% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM-specifications.R
\name{rm_km_feature}
\alias{rm_km_feature}
\title{Remove KM feature specifications from an \link{MDB} object}
\usage{
rm_km_feature(x, kmr, table, feature)
}
\arguments{
\item{x}{an \link{MDB} object to update with specification tables}

\item{kmr}{an \link{MDB} object with KM requirements}

\item{table}{the name of an existing table in x}

\item{feature}{the name of a feature with specification in x table}
}
\value{
An \link{MDB} object with relevant KM table feature specification removed
}
\description{
Remove KM feature specifications from an \link{MDB} object
}
