% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM-specifications.R
\name{rm_km_table}
\alias{rm_km_table}
\title{Remove KM table specifications from an \link{MDB} object}
\usage{
rm_km_table(x, kmr, table)
}
\arguments{
\item{x}{an \link{MDB} object to update with specification tables}

\item{kmr}{an \link{MDB} object with KM requirements}

\item{table}{the name of an existing table in x specification}
}
\value{
An \link{MDB} object with relevant KM table specification removed
}
\description{
Remove KM table specifications from an \link{MDB} object
}
