% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMDBToChTKCat.R
\name{rmMDBFromChTKCat}
\alias{rmMDBFromChTKCat}
\title{Remove an MDB (modeled database) from a \link{chTKCat} database}
\usage{
rmMDBFromChTKCat(tkcon, name)
}
\arguments{
\item{tkcon}{a \link{chTKCat} object}

\item{name}{the name of the MDB to remove}
}
\description{
Remove an MDB (modeled database) from a \link{chTKCat} database
}
