% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKCat.R, R/chTKCat.R, R/generics.R
\name{search_MDB_fields.TKCat}
\alias{search_MDB_fields.TKCat}
\alias{search_MDB_fields.chTKCat}
\alias{search_MDB_fields}
\title{Search fields in a \link{TKCat} related object}
\usage{
\method{search_MDB_fields}{TKCat}(x, searchTerm)

\method{search_MDB_fields}{chTKCat}(x, searchTerm)

search_MDB_fields(x, searchTerm)
}
\arguments{
\item{x}{a \link{TKCat} related object (e.g. \link{chTKCat})}

\item{searchTerm}{a single character with the term to search}
}
\value{
An \link{MDB} object
}
\description{
Search fields in a \link{TKCat} related object
}
