% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{set_chMDB_access}
\alias{set_chMDB_access}
\title{Set  chMDB access}
\usage{
set_chMDB_access(x, mdb, public)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{mdb}{name of the modeled database}

\item{public}{if access is public}
}
\value{
No return value, called for side effects
}
\description{
Set  chMDB access
}
