% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{set_chMDB_timestamp}
\alias{set_chMDB_timestamp}
\title{Set timestamp of the current version of an MDB in \link{chTKCat}}
\usage{
set_chMDB_timestamp(x, name, timestamp)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the database to affect}

\item{timestamp}{a single POSIXct value as a timestamp for
the chMDB instance.}
}
\value{
No return value, called for side effects
}
\description{
Set timestamp of the current version of an MDB in \link{chTKCat}
}
