% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{slice.fileMDB}
\alias{slice.fileMDB}
\title{Subset a \link{fileMDB} object according to row position in one table
and return a \link{memoMDB}}
\usage{
\method{slice}{fileMDB}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{a \link{fileMDB} object}

\item{...}{a single argument. The name of this argument should be a table
name of x and the value of this argument should be vector of integers
corresponding to row indexes.}

\item{.preserve}{not used}
}
\value{
a \link{memoMDB} object
}
\description{
Subset a \link{fileMDB} object according to row position in one table
and return a \link{memoMDB}
}
