% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memeMDB.R
\name{slice.memeMDB}
\alias{slice.memeMDB}
\title{Subset a \link{memeMDB} object according to row position in one table}
\usage{
\method{slice}{memeMDB}(x, ..., .preserve = FALSE)
}
\arguments{
\item{x}{a \link{memeMDB} object}

\item{...}{a single argument. The name of this argument should be a table
name of x and the value of this argument should be vector of integers
corresponding to row indexes.}

\item{.preserve}{not used}
}
\value{
a \link{memeMDB} object
}
\description{
Subset a \link{memeMDB} object according to row position in one table
}
\examples{
slice(hpo, HPO_diseases=1:10)

}
