% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoMDB.R
\name{slice.memoMDB}
\alias{slice.memoMDB}
\title{Subset a \link{memoMDB} object according to row position in one table}
\usage{
\method{slice}{memoMDB}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{a \link{memoMDB} object}

\item{...}{a single argument. The name of this argument should be a table
name of x and the value of this argument should be vector of integers
corresponding to row indexes.}

\item{.preserve}{not used}
}
\value{
a \link{memoMDB} object
}
\description{
Subset a \link{memoMDB} object according to row position in one table
}
