% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{[.chMDB}
\alias{[.chMDB}
\title{Load a subset of tables from a \link{chMDB} object as an \link{internalMDB} object}
\usage{
\method{[}{chMDB}(x, i = NULL)
}
\arguments{
\item{i}{the name of the tables to be taken. If NULL (default), all
the tables are loaded}
}
\description{
Load a subset of tables from a \link{chMDB} object as an \link{internalMDB} object
}
