% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKCat.R
\name{TKCat}
\alias{TKCat}
\alias{names<-.TKCat}
\alias{[.TKCat}
\alias{c.TKCat}
\title{TKCat: a catalog of \link{MDB}}
\usage{
TKCat(..., list = NULL)

\method{names}{TKCat}(x) <- value

\method{[}{TKCat}(x, i)

\method{c}{TKCat}(...)
}
\arguments{
\item{...}{\link{TKCat} objects}

\item{list}{a list of \link{MDB} objects}

\item{x}{a \link{TKCat} object}

\item{value}{new \link{MDB} names}

\item{i}{index or names of the MDB to take}
}
\value{
a TKCat object
}
\description{
TKCat: a catalog of \link{MDB}
}
\seealso{
\link{scan_fileMDBs}
}
