% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{unarchive_chMDB}
\alias{unarchive_chMDB}
\title{Unarchive a chMDB in a \link{chTKCat}}
\usage{
unarchive_chMDB(x, name)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the database to archive}
}
\value{
No return value, called for side effects
}
\description{
Unarchive a chMDB in a \link{chTKCat}
}
