% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{update_chMDB_grants}
\alias{update_chMDB_grants}
\title{Update grants on tables in an MDB of a \link{chTKCat} object}
\usage{
update_chMDB_grants(x, mdb)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{mdb}{name of the modeled database}
}
\value{
No return value, called for side effects
}
\description{
The update is done automatically based on user access.
}
