% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_members.R
\name{write_collection_members}
\alias{write_collection_members}
\title{Write a collection member JSON file}
\usage{
write_collection_members(colMembers, path = NA, collection = NULL)
}
\arguments{
\item{colMembers}{A tibble as returned by \code{\link[=read_collection_members]{read_collection_members()}}}

\item{path}{the JSON file to write. If NA (default), the JSON file is
not written but returned by the function.}

\item{collection}{The collection definition (json string).
If NULL (default), it is taken from TKCat environment
(see \code{\link[=list_local_collections]{list_local_collections()}}.}
}
\description{
Write a collection member JSON file
}
