% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R
\name{write_MergeTree}
\alias{write_MergeTree}
\title{Write a Clickhouse
\href{https://clickhouse.com/docs/en/engines/table-engines/mergetree-family/mergetree/}{MergeTree}
table}
\usage{
write_MergeTree(
  con,
  dbName,
  tableName,
  value,
  rtypes = NULL,
  nullable = NULL,
  sortKey = NULL
)
}
\arguments{
\item{con}{the clickhouse connection}

\item{dbName}{the name of the database}

\item{tableName}{the name of the table}

\item{value}{the table to import}

\item{rtypes}{a named character vector giving the R type of each and every
columns. If NULL (default), types are guessed from value.}

\item{nullable}{a character vector indicating the name of the columns
which are nullable (default: NULL)}

\item{sortKey}{a character vector indicating the name of the columns
used in the sort key. If NULL (default), all the non-nullable columns
are used in the key.}
}
\value{
No return value, called for side effects
}
\description{
Write a Clickhouse
\href{https://clickhouse.com/docs/en/engines/table-engines/mergetree-family/mergetree/}{MergeTree}
table
}
