% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMDBToChTKCat.R
\name{writeMergeTree}
\alias{writeMergeTree}
\title{Write a Clickhouse
\href{https://clickhouse.yandex/docs/en/operations/table_engines/mergetree/}{MergeTree}
table}
\usage{
writeMergeTree(
  conn,
  dbName,
  tableName,
  value,
  rtypes,
  nullable = NULL,
  indexes = NULL
)
}
\arguments{
\item{conn}{the clickhouse connection}

\item{dbName}{the name of the database}

\item{tableName}{the name of the table}

\item{value}{the table to import}

\item{rtypes}{a named character vector giving the R type of each and every
columns}

\item{nullable}{a character vector indicating the name of the columns
which are nullable (default: NULL)}

\item{indexes}{a character vector indicating the name of the columns
which are indexes (default: NULL)}
}
\description{
Write a Clickhouse
\href{https://clickhouse.yandex/docs/en/operations/table_engines/mergetree/}{MergeTree}
table
}
