% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{add_chTKCat_collection}
\alias{add_chTKCat_collection}
\title{Import a collection in a \link{chTKCat} database}
\usage{
add_chTKCat_collection(x, json, overwrite = FALSE)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{json}{a single character indicating the collection to import. Can be:
\itemize{
\item a path to a file
\item the name of a local collection (see \code{\link[=list_local_collections]{list_local_collections()}})
\item the json text defining the collection
}}

\item{overwrite}{a logical indicating if the existing collection should
be replaced.}
}
\value{
No return value, called for side effects
}
\description{
Import a collection in a \link{chTKCat} database
}
