% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/manage_kmr.R, R/manage_kmspec.R
\name{add_helpers}
\alias{add_helpers}
\alias{add_helpers.KMR}
\alias{add_helpers.MDB}
\title{Add a set of helper functions to a compatible object}
\usage{
add_helpers(x, code, name, language, ...)

\method{add_helpers}{KMR}(x, code, name, language, ...)

\method{add_helpers}{MDB}(x, code, name, language, kmr, ...)
}
\arguments{
\item{x}{an object accepting helpers}

\item{code}{file path to the source code of helper functions}

\item{name}{the name of the helper set}

\item{language}{the programming language of the code (default: "R")}

\item{...}{method specific parameters}

\item{kmr}{an \link{MDB} object with KM requirements}
}
\value{
Return x with additional helpers
}
\description{
Add a set of helper functions to a compatible object
}
