% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{as_chMDB}
\alias{as_chMDB}
\title{Push an \link{MDB} object in a ClickHouse database}
\usage{
as_chMDB(x, tkcon, overwrite = FALSE, by = 10^5)
}
\arguments{
\item{x}{an \link{MDB} object}

\item{tkcon}{a \link{chTKCat} object}

\item{overwrite}{a logical indicating if existing data should be overwritten
(default: FALSE)}

\item{by}{the size of the batch: number of records to write
together (default: 10^5)}
}
\value{
A \link{chMDB} object.
}
\description{
Push an \link{MDB} object in a ClickHouse database
}
