% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{check_chTKCat}
\alias{check_chTKCat}
\title{Check a \link{chTKCat} object}
\usage{
check_chTKCat(x, verbose = FALSE)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{verbose}{a logical indicating if information messages should be
displayed.}
}
\value{
Invisible result: \link{chTKCat} object
}
\description{
Check a \link{chTKCat} object
}
