% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{checkTable}
\alias{checkTable}
\title{Check table according to a data model}
\usage{
checkTable(x, tableModel)
}
\arguments{
\item{x}{a data.frame}

\item{tableModel}{a \link[ReDaMoR:RelTableModel]{ReDaMoR::RelTableModel} object}
}
\value{
Nothing. The function throws an error if check fails.
}
\description{
Check table according to a data model
}
