% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{chMDB}
\alias{chMDB}
\title{Get a ClickHouse modeled database}
\usage{
chMDB(tkcon, dbName)
}
\arguments{
\item{tkcon}{a chTKCat object}

\item{dbName}{the name of the database}
}
\value{
A chMDB object
}
\description{
Get a ClickHouse modeled database
}
