% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memoMDB.R
\name{collection_members<-.fileMDB}
\alias{collection_members<-.fileMDB}
\alias{collection_members<-}
\alias{collection_members<-.memoMDB}
\title{Set collection members}
\usage{
\method{collection_members}{fileMDB}(x) <- value

collection_members(x) <- value

\method{collection_members}{memoMDB}(x) <- value
}
\arguments{
\item{x}{a \link{memoMDB} object}

\item{value}{A data.frame with the following columns:
\itemize{
\item \strong{collection} (character): The name of the collection
\item \strong{cid} (character): Collection identifier
\item \strong{resource} (character): The name of the resource
\item \strong{mid} (integer): The identifier of the member
\item \strong{table} (character): The table recording collection information
\item \strong{field} (character): The collection field.
\item \strong{static} (logical): TRUE if the field value is common to all elements.
\item \strong{value} (character): The name of the table column if static is FALSE
or the field value if static is TRUE.
\item \strong{type} (character): the type of the field.
(not necessarily used ==> NA if not)
}}
}
\description{
Set collection members
}
