% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDB.R, R/generics.R
\name{count_records.MDB}
\alias{count_records.MDB}
\alias{count_records}
\title{Count the number of records}
\usage{
\method{count_records}{MDB}(x, ...)

count_records(x, ...)
}
\arguments{
\item{x}{an object with embedded data tables}

\item{...}{the name of the tables to consider (default: all of them)}
}
\value{
A named vector with the number of records per table.
}
\description{
Count the number of records
}
