% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/fileMDB.R, R/generics.R,
%   R/memoMDB.R, R/metaMDB.R
\name{count_records.chMDB}
\alias{count_records.chMDB}
\alias{count_records.fileMDB}
\alias{count_records}
\alias{count_records.memoMDB}
\alias{count_records.metaMDB}
\title{Count the number of records}
\usage{
\method{count_records}{chMDB}(x, ...)

\method{count_records}{fileMDB}(x, ...)

count_records(x, ...)

\method{count_records}{memoMDB}(x, ...)

\method{count_records}{metaMDB}(x, ...)
}
\arguments{
\item{x}{an object with embedded data tables}

\item{...}{the name of the tables to consider (default: all of them)}
}
\value{
A named vector with the number of records per table.
}
\description{
Count the number of records
}
