% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{create_chMDB}
\alias{create_chMDB}
\title{Create a database in a \link{chTKCat}}
\usage{
create_chMDB(x, name, public = FALSE)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the new database}

\item{public}{if the database data are accessible to any user (default:FALSE)}
}
\value{
No return value, called for side effects
}
\description{
Create a database in a \link{chTKCat}
}
