% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{create_chTKCat_user}
\alias{create_chTKCat_user}
\title{Create a chTKCat user}
\usage{
create_chTKCat_user(
  x,
  login,
  password,
  contact,
  admin = FALSE,
  provider = admin
)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{login}{user login}

\item{password}{user password (NA ==> no password)}

\item{contact}{contact information (can be NA)}

\item{admin}{a logical indicating if the user is an admin of the chTKCat
instance (default: TRUE)}

\item{provider}{a logical indicating if the user is data provider (TRUE)
or a data consumer (FALSE: default). If admin
is set to TRUE provider will be set to TRUE}
}
\value{
No return value, called for side effects
}
\description{
Create a chTKCat user
}
