% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{data_model}
\alias{data_model}
\title{Get object data model}
\usage{
data_model(x, ...)
}
\arguments{
\item{x}{an object with an embedded data model}

\item{...}{further arguments for the method}
}
\description{
Get object data model
}
