% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memoMDB.R,
%   R/metaMDB.R
\name{data_model.fileMDB}
\alias{data_model.fileMDB}
\alias{data_model}
\alias{data_model.memoMDB}
\alias{data_model.metaMDB}
\title{Get object data model}
\usage{
\method{data_model}{fileMDB}(x, ...)

data_model(x, ...)

\method{data_model}{memoMDB}(x, ...)

\method{data_model}{metaMDB}(x, rtOnly = FALSE, ...)
}
\arguments{
\item{x}{a \link{metaMDB} object}

\item{...}{not used}

\item{rtOnly}{if TRUE, the function only returns the relationTables and
the foreign tables (default: FALSE)}
}
\value{
A \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object
}
\description{
Get object data model
}
