% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memeMDB.R
\name{data_model.fileMDB}
\alias{data_model.fileMDB}
\alias{data_model}
\alias{data_model.memeMDB}
\title{Get object data model}
\usage{
\method{data_model}{fileMDB}(x, ...)

data_model(x, ...)

\method{data_model}{memeMDB}(x, ...)
}
\arguments{
\item{x}{a \link{memeMDB} object}

\item{...}{not used}
}
\value{
A \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object
}
\description{
Get object data model
}
