% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/fileMDB.R, R/generics.R,
%   R/memoMDB.R, R/metaMDB.R
\name{data_model.chMDB}
\alias{data_model.chMDB}
\alias{data_model.fileMDB}
\alias{data_model}
\alias{data_model.memoMDB}
\alias{data_model.metaMDB}
\title{Get object data model}
\usage{
\method{data_model}{chMDB}(x, ...)

\method{data_model}{fileMDB}(x, ...)

data_model(x, ...)

\method{data_model}{memoMDB}(x, ...)

\method{data_model}{metaMDB}(x, rtOnly = FALSE, recursive = FALSE, ...)
}
\arguments{
\item{x}{an object with an embedded data model}

\item{...}{method specific parameters}

\item{rtOnly}{if TRUE, the function only returns the relational tables
and the corresponding foreign tables (default: FALSE)}

\item{recursive}{if TRUE and rtOnly, the function returns also the
relational tables from embedded metaMDBs.}
}
\value{
A \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object
}
\description{
Get object data model
}
