% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memeMDB.R
\name{data_tables.fileMDB}
\alias{data_tables.fileMDB}
\alias{data_tables}
\alias{data_tables.memeMDB}
\title{Get object data tables}
\usage{
\method{data_tables}{fileMDB}(x, ...)

data_tables(x, ...)

\method{data_tables}{memeMDB}(x, ...)
}
\arguments{
\item{x}{a \link{memeMDB} object}

\item{...}{the name of the tables to get (default: all of them)}
}
\value{
A list of \link[dplyr:reexports]{dplyr::tibble}
}
\description{
Get object data tables
}
