% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/fileMDB.R, R/generics.R,
%   R/memoMDB.R, R/metaMDB.R
\name{data_tables.chMDB}
\alias{data_tables.chMDB}
\alias{data_tables.fileMDB}
\alias{data_tables}
\alias{data_tables.memoMDB}
\alias{data_tables.metaMDB}
\title{Get object data tables}
\usage{
\method{data_tables}{chMDB}(x, ..., skip = 0, n_max = Inf)

\method{data_tables}{fileMDB}(x, ..., skip = 0, n_max = Inf)

data_tables(x, ..., skip = 0, n_max = Inf)

\method{data_tables}{memoMDB}(x, ..., skip = 0, n_max = Inf)

\method{data_tables}{metaMDB}(x, ...)
}
\arguments{
\item{x}{an object with embedded data tables}

\item{...}{the name of the tables to get (default: all of them)}

\item{skip}{the number of rows to skip (default: 0)}

\item{n_max}{maximum number of rows to return (default: Inf)}
}
\value{
A list of \link[dplyr:reexports]{dplyr::tibble}
}
\description{
Get object data tables
}
