% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memoMDB.R,
%   R/metaMDB.R
\name{db_info<-.fileMDB}
\alias{db_info<-.fileMDB}
\alias{db_info<-}
\alias{db_info<-.memoMDB}
\alias{db_info<-.metaMDB}
\title{Set DB information}
\usage{
\method{db_info}{fileMDB}(x) <- value

db_info(x) <- value

\method{db_info}{memoMDB}(x) <- value

\method{db_info}{metaMDB}(x) <- value
}
\arguments{
\item{x}{a \link{metaMDB} object}

\item{value}{a list with DB information:
"name", "title", "description", "url",
"version", "maintainer".}
}
\description{
Set DB information
}
