% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memoMDB.R
\name{db_info.fileMDB}
\alias{db_info.fileMDB}
\alias{db_info}
\alias{db_info.memoMDB}
\title{Get DB information}
\usage{
\method{db_info}{fileMDB}(x, ...)

db_info(x, ...)

\method{db_info}{memoMDB}(x, ...)
}
\arguments{
\item{x}{a \link{memoMDB} object}

\item{...}{not used}
}
\value{
A list with the following elements:
\itemize{
\item \strong{name}: a single character
\item \strong{title}: a single character
\item \strong{description}: a single character
\item \strong{url}: a single character
\item \strong{version}: a single character
\item \strong{maintainer}: a single character vector
\item \strong{size}: a numeric vector providing the size of the DB in bytes
}
}
\description{
Get DB information
}
