% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{db_info}
\alias{db_info}
\title{Get DB information}
\usage{
db_info(x, ...)
}
\arguments{
\item{x}{an object with embedded database information}

\item{...}{further arguments for the method}
}
\description{
Get DB information
}
