% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R, R/memoMDB.R,
%   R/metaMDB.R
\name{db_info.fileMDB}
\alias{db_info.fileMDB}
\alias{db_info<-.fileMDB}
\alias{db_info}
\alias{db_info<-}
\alias{db_info.memoMDB}
\alias{db_info<-.memoMDB}
\alias{db_info.metaMDB}
\alias{db_info<-.metaMDB}
\title{DB information}
\usage{
\method{db_info}{fileMDB}(x, ...)

\method{db_info}{fileMDB}(x) <- value

db_info(x, ...)

db_info(x) <- value

\method{db_info}{memoMDB}(x, ...)

\method{db_info}{memoMDB}(x) <- value

\method{db_info}{metaMDB}(x, ...)

\method{db_info}{metaMDB}(x) <- value
}
\arguments{
\item{x}{a \link{metaMDB} object}

\item{...}{not used}

\item{value}{a list with DB information:
"name", "title", "description", "url",
"version", "maintainer".}
}
\value{
A list with the following elements:
\itemize{
\item \strong{name}: a single character
\item \strong{title}: a single character
\item \strong{description}: a single character
\item \strong{url}: a single character
\item \strong{version}: a single character
\item \strong{maintainer}: a single character vector
\item \strong{size}: a numeric vector providing the size of the DB in bytes
}
}
\description{
DB information

DB information
}
