% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R, R/chMDB.R, R/chTKCat.R,
%   R/generics.R, R/metaMDB.R
\name{db_reconnect.ClickhouseConnection}
\alias{db_reconnect.ClickhouseConnection}
\alias{db_reconnect.chMDB}
\alias{db_reconnect.chTKCat}
\alias{db_reconnect}
\alias{db_reconnect.metaMDB}
\title{Reconnect an object to a database}
\usage{
\method{db_reconnect}{ClickhouseConnection}(x, user, password, ntries = 3, settings = list(), ...)

\method{db_reconnect}{chMDB}(x, user, password, ntries = 3, ...)

\method{db_reconnect}{chTKCat}(x, user, password, ntries = 3, ...)

db_reconnect(x, user, password, ntries = 3, ...)

\method{db_reconnect}{metaMDB}(x, user, password, ntries = 3, ...)
}
\arguments{
\item{x}{an object with a database connection}

\item{user}{user name. If not provided, it's taken from x}

\item{password}{user password. If not provided, first the function
tries to connect without any password.If it fails, the function asks the
user to provide a password.}

\item{ntries}{the number of times the user can enter a wrong password
(default: 3)}

\item{settings}{list of
\href{https://clickhouse.com/docs/en/operations/settings/settings/}{Clickhouse settings}}

\item{...}{additional parameters for methods}
}
\value{
A new database connection object.
}
\description{
Reconnect an object to a database
}
