% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{db_tables}
\alias{db_tables}
\title{Get the DB tables from a \link{chMDB} object}
\usage{
db_tables(x)
}
\arguments{
\item{x}{a \link{chMDB} object}
}
\value{
a list with a chTKCat object (tkcon) and
a named vector of DB table names (dbTables).
}
\description{
Get the DB tables from a \link{chMDB} object
}
