% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{dbInfo.internalMDB}
\alias{dbInfo.internalMDB}
\title{Get DB information of an \link{internalMDB} object}
\usage{
\method{dbInfo}{internalMDB}(x, countRecords = TRUE)
}
\arguments{
\item{x}{an \link{internalMDB} object}

\item{countRecords}{a single logical. If TRUE (default), the number of
records is also returned.}
}
\value{
A list with the following elements:
\itemize{
\item \strong{name}: a single character
\item \strong{title}: a single character
\item \strong{description}: a single character
\item \strong{url}: a single character
\item \strong{version}: a single character
\item (\strong{table_records}): a numeric giving the number of records per table
\item (\strong{records}): a single numeric
}
}
\description{
Get DB information of an \link{internalMDB} object
}
