% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{disconnect_chTKCat}
\alias{disconnect_chTKCat}
\title{Disconnect from a ClickHouse TKCat instance}
\usage{
disconnect_chTKCat(x)
}
\arguments{
\item{x}{a chTKCat object}
}
\description{
Disconnect from a ClickHouse TKCat instance
}
\seealso{
\code{\link{chTKCat}}
}
