% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{empty_chMDB}
\alias{empty_chMDB}
\title{Empty a chMDB in a \link{chTKCat}}
\usage{
empty_chMDB(x, name)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the database to empty}
}
\value{
No return value, called for side effects
}
\description{
Empty a chMDB in a \link{chTKCat}
}
