% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R
\name{fileMDB}
\alias{fileMDB}
\title{Create an MDB (Modeled DataBase) based on files}
\usage{
fileMDB(
  dataFiles,
  dbInfo,
  dataModel,
  readParameters = list(),
  collectionMembers = NULL,
  n_max = 10,
  verbose = FALSE
)
}
\arguments{
\item{dataFiles}{a named vector of path to data files with
\code{all(names(dataFiles) \%in\% names(dataModel))}}

\item{dbInfo}{a list with DB information:
"name", "title", "description", "url",
"version", "maintainer".}

\item{dataModel}{a \link[ReDaMoR:RelDataModel]{ReDaMoR::RelDataModel} object}

\item{readParameters}{a list of parameters for reading the data file.
(e.g. \code{list(delim='\\t', quoted_na=FALSE,)})}

\item{collectionMembers}{the members of collections as provided to the
\link{collection_members<-.fileMDB} function (default: NULL ==> no member).}

\item{n_max}{maximum number of records to read
for checks purpose (default: 10). See also \code{\link[ReDaMoR:confront_data]{ReDaMoR::confront_data()}}.}

\item{verbose}{if TRUE display the data confrontation report}
}
\value{
A \link{fileMDB} object
}
\description{
Create an MDB (Modeled DataBase) based on files
}
