% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{filter.chMDB}
\alias{filter.chMDB}
\title{Filter a \link{chMDB} object and return a \link{memoMDB}}
\usage{
\method{filter}{chMDB}(.data, ..., by = 10^5, .preserve = FALSE)
}
\arguments{
\item{.data}{a \link{chMDB} object}

\item{...}{each argument should have the name of one of the tables of the
\link{chMDB} object and contain a simple logical expression involving
the names of the corresponding table.}

\item{by}{the size of the batch: number of records to filter
together (default: 10^5)}

\item{.preserve}{not used}
}
\value{
a \link{memoMDB} object
}
\description{
Filter a \link{chMDB} object and return a \link{memoMDB}
}
