% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filter_with_tables}
\alias{filter_with_tables}
\title{Filter an MDB object according to provided tables}
\usage{
filter_with_tables(x, tables, checkTables = TRUE)
}
\arguments{
\item{x}{an MDB object}

\item{tables}{a named list of tibbles to filter with. The names should
correspond to the table names in x and the tibbles should fit the
data model.}

\item{checkTables}{if TRUE, the tables are confronted to their model
in the data model of x.}
}
\value{
a \link{memoMDB} object
}
\description{
Filter an MDB object according to provided tables
}
