% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{format.chTKCat}
\alias{format.chTKCat}
\title{Format a \link{chTKCat} object for printing}
\usage{
\method{format}{chTKCat}(x, ...)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{...}{not used}
}
\value{
A single character
}
\description{
Format a \link{chTKCat} object for printing
}
