% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{get_chMDB_timestamps}
\alias{get_chMDB_timestamps}
\title{Get instance timestamps of an MDB in \link{chTKCat}}
\usage{
get_chMDB_timestamps(x, name)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{name}{the name of the database}
}
\value{
A tibble with the instance "timestamp" and a logical indicating if
it's the "current" one or not.
}
\description{
Get instance timestamps of an MDB in \link{chTKCat}
}
