% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{get_chTKCat_collection}
\alias{get_chTKCat_collection}
\title{Get a collection from a \link{chTKCat}}
\usage{
get_chTKCat_collection(x, title)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{title}{the title of the collection to get}
}
\value{
The definition of the collection as a JSON string.
}
\description{
Get a collection from a \link{chTKCat}
}
