% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge-access.R
\name{get_pok}
\alias{get_pok}
\title{Get a piece of knowledge from a \link{chTKCat} connection}
\usage{
get_pok(x, mdb_name, corpus_name)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{mdb_name}{the piece of knowlege to get}

\item{corpus_name}{the corpus to which the piece belongs}
}
\description{
Get a piece of knowledge from a \link{chTKCat} connection
}
