% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knowledge-access.R
\name{get_pok}
\alias{get_pok}
\title{Get a piece of knowledge from a \link{chTKCat} connection or a \link{TKCat} object}
\usage{
get_pok(tkcat, mdb_name, corpus_name)
}
\arguments{
\item{tkcat}{a \link{chTKCat} or a \link{TKCat} object}

\item{mdb_name}{the piece of knowledge to get}

\item{corpus_name}{the corpus to which the piece belongs}
}
\description{
Get a piece of knowledge from a \link{chTKCat} connection or a \link{TKCat} object
}
