% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POK.R
\name{get_POK}
\alias{get_POK}
\title{Get a \link{POK} from a \link{chTKCat} connection or a \link{TKCat} object}
\usage{
get_POK(x, mdb, kmr)
}
\arguments{
\item{x}{a \link{chTKCat} or a \link{TKCat} object}

\item{mdb}{\link{MDB} object with KM specifications or its name in tkcat}

\item{kmr}{\link{KMR} object with KM requirements or its name in tkcat}
}
\value{
A \link{POK} object
}
\description{
Get a \link{POK} from a \link{chTKCat} connection or a \link{TKCat} object
}
