% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R, R/chTKCat.R, R/generics.R
\name{get_query.chMDB}
\alias{get_query.chMDB}
\alias{get_query.chTKCat}
\alias{get_query}
\title{Get SQL query}
\usage{
\method{get_query}{chMDB}(x, query, autoalias = !is_current_chMDB(x), ...)

\method{get_query}{chTKCat}(x, query, ...)

get_query(x, query, ...)
}
\arguments{
\item{x}{an object with a database connection}

\item{query}{the SQL query}

\item{autoalias}{Change this parameter only if you know what you're doing.
if TRUE, make relevant alias to query the chMDB
using the table names from the data model. If FALSE, the user must know the
table instance name in the remote database. By default, autoalias is set
to TRUE when using a non-current instance of the database.}

\item{...}{method specific parameters}
}
\value{
A tibble with query results
}
\description{
Get SQL query
}
