% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/manage_kmr.R, R/manage_kmspec.R
\name{get_R_helpers}
\alias{get_R_helpers}
\alias{get_R_helpers.KMR}
\alias{get_R_helpers.MDB}
\title{Get a set of helper functions from an object}
\usage{
get_R_helpers(x, hnames, ...)

\method{get_R_helpers}{KMR}(x, hnames = NA, tkcat = NULL, mdb = NULL, ...)

\method{get_R_helpers}{MDB}(x, hnames = NA, kmr, ...)
}
\arguments{
\item{x}{an object with helpers}

\item{hnames}{the names of the helper sets. If NA (default),
all available are sourced.}

\item{...}{method specific parameters}

\item{tkcat}{A \link{TKCat} or \link{chTKCat} object to make available in
helper environment}

\item{mdb}{An \link{MDB} object to make available in helper environment}

\item{kmr}{an \link{MDB} object with KM requirements}
}
\value{
Return a list of functions
}
\description{
Get a set of helper functions from an object
}
\details{
x and mdb objects are made available in helpers environment as
'THISKMR' and 'THISMDB' objects respectively and can be used as such within
helpers code.

x and kmr objects are made available in helpers environment as
'THISMDB' and 'THISKMR' objects respectively and can be used as such within
helpers code.
}
