% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KM-specifications.R, R/KMR.R, R/generics.R
\name{get_R_helpers.MDB}
\alias{get_R_helpers.MDB}
\alias{get_R_helpers.KMR}
\alias{get_R_helpers}
\title{Get a set of helper functions from an object}
\usage{
\method{get_R_helpers}{MDB}(x, hnames = NA, kmr, tkcat = NULL, ...)

\method{get_R_helpers}{KMR}(x, hnames = NA, tkcat = NULL, mdb = NULL, ...)

get_R_helpers(x, hnames, ...)
}
\arguments{
\item{x}{an object with helpers}

\item{hnames}{the names of the helper sets. If NA (default),
all available are sourced.}

\item{kmr}{a \link{KMR} object}

\item{tkcat}{A \link{TKCat} or \link{chTKCat} object to make available in
helper environment}

\item{...}{method specific parameters}

\item{mdb}{An \link{MDB} object to make available in helper environment}
}
\value{
Return a list of functions
}
\description{
Get a set of helper functions from an object
}
\details{
x, kmr and tkcat objects are made available in helpers environment
as 'THISMDB', 'THISKMR' and 'THISTKCAT' objects respectively and can be used
as such within helpers code.

x, tkcat and mdb objects are made available in helpers environment
as 'THISKMR', 'THISTKCAT' and 'THISMDB' objects respectively and can be used
as such within helpers code.
}
