% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importLocalCollection.R
\name{importLocalCollection}
\alias{importLocalCollection}
\title{Import a collection definition in the local environment}
\usage{
importLocalCollection(f, overwrite = FALSE)
}
\arguments{
\item{f}{the JSON file to import}

\item{overwrite}{a single logical. If TRUE the collection is overwritten
if it already exists (default: FALSE)}
}
\description{
Import a collection definition in the local environment
}
