% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{init_chTKCat}
\alias{init_chTKCat}
\title{Initialize a chTKCat database}
\usage{
init_chTKCat(x, instance, version)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{instance}{instance name of the database}

\item{version}{version name of the database}
}
\value{
a \link{chTKCat}
}
\description{
Initialize a chTKCat database
}
