% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalMDB.R
\name{internalMDB}
\alias{internalMDB}
\title{Create an internal database: a set of tables with a data model}
\usage{
internalMDB(dataModel, dbTables, checkTables = TRUE)
}
\arguments{
\item{dataModel}{a RelDataModel object}

\item{checkTables}{a single logical. If TRUE (default), tables are confronted
to the dataModel.}

\item{dbTable}{a list of tables}
}
\value{
An internalMDB object
}
\description{
Create an internal database: a set of tables with a data model
}
