% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chMDB.R
\name{is.chMDB}
\alias{is.chMDB}
\title{Check if the object is a \link{chMDB} object}
\usage{
is.chMDB(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is a \link{chMDB} object
}
\description{
Check if the object is a \link{chMDB} object
}
