% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDB.R
\name{is.MDB}
\alias{is.MDB}
\title{Check if the object is an \link{MDB} object}
\usage{
is.MDB(x)
}
\arguments{
\item{x}{any object}
}
\value{
A single logical: TRUE if x is an MDB object.
}
\description{
Check if the object is an \link{MDB} object
}
