% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{list_ChTKCat_collections}
\alias{list_ChTKCat_collections}
\title{List collections available in a \link{chTKCat}}
\usage{
list_ChTKCat_collections(tkcon)
}
\arguments{
\item{tkcon}{a \link{chTKCat} object}
}
\description{
List collections available in a \link{chTKCat}
}
