% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{list_chTKCat_collections}
\alias{list_chTKCat_collections}
\title{List collections available in a \link{chTKCat}}
\usage{
list_chTKCat_collections(x, withJson = FALSE)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{withJson}{if TRUE, returns the json strings of the collection
(default: FALSE)}
}
\value{
A tibble with the title, the description and optionally the json
definition of the collections
}
\description{
List collections available in a \link{chTKCat}
}
