% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClickHouse.R, R/chTKCat.R, R/generics.R
\name{list_tables.DBIConnection}
\alias{list_tables.DBIConnection}
\alias{list_tables.chTKCat}
\alias{list_tables}
\title{List tables in a clickhouse database}
\usage{
\method{list_tables}{DBIConnection}(x, dbNames = NULL, ...)

\method{list_tables}{chTKCat}(x, dbNames = NULL, ...)

list_tables(x, ...)
}
\arguments{
\item{x}{an object with a clickhouse connection}

\item{dbNames}{the name of databases to focus on (default NULL ==> all)}

\item{...}{method specific parameters}
}
\value{
A tibble with at least the following columns:
\itemize{
\item \strong{database}: the name of the database
\item \strong{name}: the name of the table
\item \strong{total_rows}: the number of rows in the table
\item \strong{total_bytes}: the size of the table
\item \strong{total_columns}: the number of columns in the table
}
}
\description{
List tables in a clickhouse database
}
