% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_members.R
\name{read_collection_members}
\alias{read_collection_members}
\title{Read a collection member JSON file}
\usage{
read_collection_members(txt)
}
\arguments{
\item{txt}{a JSON string or file}
}
\value{
A tibble with the description of the collection members of a
resource
}
\description{
Read a collection member JSON file
}
