% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInternalMDB.R
\name{readInternalMDB}
\alias{readInternalMDB}
\title{Read files and create an internal database with a documented data model}
\usage{
readInternalMDB(dataModel, directory, ext = "txt", delim = "\\t",
  quoted_na = FALSE, ...)
}
\arguments{
\item{dataModel}{a \code{\link{RelDataModel}} object}

\item{directory}{a path to the directory where to read the data}

\item{ext}{the file extension to consider (default: "ext"),}

\item{delim}{delimiter (default: '\\t')}

\item{quoted_na}{Should missing values inside quotes be treated
as missing values or as strings (the default).}

\item{...}{additional  parameters for the \code{\link{read_delim}} function}
}
\value{
An \code{\link{internalMDB}} object
}
\description{
Read files and create an internal database with a documented data model
}
