% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chTKCat.R
\name{remove_chTKCat_collection}
\alias{remove_chTKCat_collection}
\title{Remove a collection from a \link{chTKCat} database}
\usage{
remove_chTKCat_collection(x, title)
}
\arguments{
\item{x}{a \link{chTKCat} object}

\item{title}{the title of the collection to remove}
}
\value{
No return value, called for side effects
}
\description{
Remove a collection from a \link{chTKCat} database
}
