% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileMDB.R, R/generics.R
\name{write_MDB.fileMDB}
\alias{write_MDB.fileMDB}
\alias{write_MDB}
\title{Write an MDB object}
\usage{
\method{write_MDB}{fileMDB}(x, path, ...)

write_MDB(x, path, ...)
}
\arguments{
\item{x}{an MDB object}

\item{path}{the path where the MDB should be written}

\item{...}{not used}
}
\value{
A \link{fileMDB} object.
}
\description{
Write an MDB object
}
